@@ -1,31 +1,38 @@
-import type { DeployFunction } from 'hardhat-deploy/types.js'
+import { execute, artifacts } from '@rocketh'
 import type { Address } from 'viem'
 
-const func: DeployFunction = async function (hre) {
-  const { network, viem } = hre
+export default execute(
+  async ({ deploy, namedAccounts, network }) => {
+    const { deployer } = namedAccounts
 
-  let oracleAddress: Address = '0x5f4eC3Df9cbd43714FE2740f5E3616155c5b8419'
-  if (network.name !== 'mainnet') {
-    const dummyOracle = await viem.deploy('DummyOracle', [160000000000n])
-    oracleAddress = dummyOracle.address
-  }
+    let oracleAddress: Address = '0x5f4eC3Df9cbd43714FE2740f5E3616155c5b8419'
+    if (network.name !== 'mainnet') {
+      const dummyOracle = await deploy('DummyOracle', {
+        account: deployer,
+        artifact: artifacts.DummyOracle,
+        args: [160000000000n],
+      })
+      oracleAddress = dummyOracle.address
+    }
 
-  await viem.deploy('ExponentialPremiumPriceOracle', [
-    oracleAddress,
-    [0n, 0n, 20294266869609n, 5073566717402n, 158548959919n],
-    100000000000000000000000000n,
-    21n,
-  ])
-
-  return true
-}
-
-func.id = 'ExponentialPremiumPriceOracle v1.0.0'
-func.tags = [
-  'category:ethregistrar',
-  'ExponentialPremiumPriceOracle',
-  'DummyOracle',
-]
-func.dependencies = []
-
-export default func
+    await deploy('ExponentialPremiumPriceOracle', {
+      account: deployer,
+      artifact: artifacts.ExponentialPremiumPriceOracle,
+      args: [
+        oracleAddress,
+        [0n, 0n, 20294266869609n, 5073566717402n, 158548959919n],
+        100000000000000000000000000n,
+        21n,
+      ],
+    })
+  },
+  {
+    id: 'ExponentialPremiumPriceOracle v1.0.0',
+    tags: [
+      'category:ethregistrar',
+      'ExponentialPremiumPriceOracle',
+      'DummyOracle',
+    ],
+    dependencies: [],
+  },
+)
