@@ -1,23 +1,26 @@
-import type { DeployFunction } from 'hardhat-deploy/types.js'
+import { artifacts, execute } from '@rocketh'
+import type { Artifact } from 'rocketh'
+import LegacyPublicResolverArtifact from '../../deployments/archive/PublicResolver_mainnet_9412610.sol/PublicResolver_mainnet_9412610.json'
 
-const func: DeployFunction = async function (hre) {
-  const { deployments, network, viem } = hre
+export default execute(
+  async ({ deploy, get, namedAccounts, network }) => {
+    const { deployer } = namedAccounts
 
-  const registry = await viem.getContract('ENSRegistry')
+    if (!network.tags?.legacy) {
+      return
+    }
 
-  if (!network.tags.legacy) {
-    return
-  }
+    const registry = get<(typeof artifacts.ENSRegistry)['abi']>('ENSRegistry')
 
-  await viem.deploy('LegacyPublicResolver', [registry.address], {
-    artifact: await deployments.getArtifact('PublicResolver_mainnet_9412610'),
-  })
-
-  return true
-}
-
-func.id = 'PublicResolver v1.0.0'
-func.tags = ['category:resolvers', 'LegacyPublicResolver']
-func.dependencies = ['ENSRegistry']
-
-export default func
+    await deploy('LegacyPublicResolver', {
+      account: deployer,
+      artifact: LegacyPublicResolverArtifact as unknown as Artifact,
+      args: [registry.address],
+    })
+  },
+  {
+    id: 'PublicResolver v1.0.0',
+    tags: ['category:resolvers', 'LegacyPublicResolver'],
+    dependencies: ['ENSRegistry'],
+  },
+)
