@@ -1,62 +1,58 @@
-import type { DeployFunction } from 'hardhat-deploy/types.js'
-import { labelhash, namehash } from 'viem'
+import { artifacts, execute } from '@rocketh'
+import { namehash } from 'viem'
 
-const func: DeployFunction = async function (hre) {
-  const { network, viem } = hre
+export default execute(
+  async ({ deploy, get, read, execute: write, namedAccounts, network }) => {
+    const { deployer, owner } = namedAccounts
 
-  const { owner } = await viem.getNamedClients()
+    const defaultReverseRegistrar = get<
+      (typeof artifacts.DefaultReverseRegistrar)['abi']
+    >('DefaultReverseRegistrar')
+    const registry = get<(typeof artifacts.ENSRegistry)['abi']>('ENSRegistry')
+    const root = get<(typeof artifacts.Root)['abi']>('Root')
 
-  const defaultReverseRegistrar = await viem.getContract(
-    'DefaultReverseRegistrar',
-  )
+    const defaultReverseResolver = await deploy('DefaultReverseResolver', {
+      account: deployer,
+      artifact: artifacts.DefaultReverseResolver,
+      args: [defaultReverseRegistrar.address],
+    })
 
-  const defaultReverseResolver = await viem.deploy('DefaultReverseResolver', [
-    defaultReverseRegistrar.address,
-  ])
+    if (network.name === 'mainnet' && !network.tags.tenderly) return
 
-  // Only attempt to make controller etc changes directly on testnets
-  if (network.name === 'mainnet' && !network.tags.tenderly) return
+    const currentRootOwner = await read(root, {
+      functionName: 'owner',
+      args: [],
+    })
+    const currentReverseOwner = await read(registry, {
+      functionName: 'owner',
+      args: [namehash('reverse')],
+    })
+    if (currentRootOwner === owner && currentReverseOwner !== owner) {
+      console.log(`  - Setting owner of .reverse to owner on root`)
+      await write(root, {
+        functionName: 'transferOwnership',
+        args: [owner],
+        account: deployer,
+      })
+    } else if (currentRootOwner !== owner) {
+      console.warn(
+        `  - WARN: Root owner account not available, skipping .reverse setup on registry`,
+      )
+      return
+    }
 
-  const registry = await viem.getContract('ENSRegistry')
-  const root = await viem.getContract('Root')
-
-  const currentRootOwner = await root.read.owner()
-  const currentReverseOwner = await registry.read.owner([namehash('reverse')])
-  if (
-    currentRootOwner === owner.address &&
-    currentReverseOwner !== owner.address
-  ) {
-    const setReverseOwnerHash = await root.write.setSubnodeOwner(
-      [labelhash('reverse'), owner.address],
-      { account: owner.account },
-    )
     console.log(
-      `Setting owner of .reverse to owner on root (tx: ${setReverseOwnerHash})...`,
+      `  - Setting resolver of .reverse to DefaultReverseResolver on registry`,
     )
-    await viem.waitForTransactionSuccess(setReverseOwnerHash)
-  } else if (currentRootOwner !== owner.address) {
-    console.warn(
-      'Root owner account not available, skipping .reverse setup on registry',
-    )
-    return
-  }
-
-  const setResolverHash = await registry.write.setResolver(
-    [namehash('reverse'), defaultReverseResolver.address],
-    {
-      account: owner.account,
-    },
-  )
-  console.log(
-    `Setting resolver of .reverse to DefaultReverseResolver on registry (tx: ${setResolverHash})...`,
-  )
-  await viem.waitForTransactionSuccess(setResolverHash)
-
-  return true
-}
-
-func.id = 'DefaultReverseResolver v1.0.0'
-func.tags = ['category:reverseresolver', 'DefaultReverseResolver']
-func.dependencies = ['ENSRegistry', 'Root', 'DefaultReverseRegistrar']
-
-export default func
+    await write(registry, {
+      functionName: 'setResolver',
+      args: [namehash('reverse'), defaultReverseResolver.address],
+      account: owner,
+    })
+  },
+  {
+    id: 'DefaultReverseResolver v1.0.0',
+    tags: ['category:reverseresolver', 'DefaultReverseResolver'],
+    dependencies: ['ENSRegistry', 'Root', 'DefaultReverseRegistrar'],
+  },
+)
