@@ -1,61 +1,66 @@
-import type { DeployFunction } from 'hardhat-deploy/types.js'
-import { zeroHash } from 'viem'
+import { artifacts, execute } from '@rocketh'
+import { getAddress, zeroHash, type Address } from 'viem'
 
-const func: DeployFunction = async function (hre) {
-  const { network, viem } = hre
+export default execute(
+  async ({ get, read, execute: write, namedAccounts, network }) => {
+    const { deployer, owner } = namedAccounts
 
-  const { deployer, owner } = await viem.getNamedClients()
+    if (!network.tags.use_root) {
+      console.warn('  - WARN: Skipping root setup (use_root not enabled)')
+      return
+    }
 
-  if (!network.tags.use_root) {
+    console.log('  - Running root setup')
+
+    const registry = get<(typeof artifacts.ENSRegistry)['abi']>('ENSRegistry')
+    const root = get<(typeof artifacts.Root)['abi']>('Root')
+
+    console.log(`  - Setting owner of root node to root contract`)
+    await write(registry, {
+      functionName: 'setOwner',
+      args: [zeroHash, root.address],
+      account: deployer,
+    })
+
+    const rootOwner = await read(root, {
+      functionName: 'owner',
+      args: [],
+    }).then((v) => getAddress(v as Address))
+
+    switch (rootOwner) {
+      case getAddress(deployer):
+        console.log(`  - Transferring ownership of root node to ${owner}`)
+        await write(root, {
+          functionName: 'transferOwnership',
+          args: [owner],
+          account: deployer,
+        })
+      case getAddress(owner):
+        const ownerIsRootController = await read(root, {
+          functionName: 'controllers',
+          args: [owner],
+        })
+        if (!ownerIsRootController) {
+          console.log(`  - Setting ${owner} as controller on root contract`)
+          await write(root, {
+            functionName: 'setController',
+            args: [owner, true],
+            account: owner,
+          })
+        }
+        break
+      default:
+        console.warn(
+          `  - WARN: Root is owned by ${rootOwner}; cannot transfer to owner account`,
+        )
+        break
+    }
+
     return true
-  }
-
-  console.log('Running root setup')
-
-  const registry = await viem.getContract('ENSRegistry')
-  const root = await viem.getContract('Root')
-
-  const setOwnerHash = await registry.write.setOwner([zeroHash, root.address])
-  console.log(
-    `Setting owner of root node to root contract (tx: ${setOwnerHash})...`,
-  )
-  await viem.waitForTransactionSuccess(setOwnerHash)
-
-  const rootOwner = await root.read.owner()
-
-  switch (rootOwner) {
-    case deployer.address:
-      const transferOwnershipHash = await root.write.transferOwnership([
-        owner.address,
-      ])
-      console.log(
-        `Transferring root ownership to final owner (tx: ${transferOwnershipHash})...`,
-      )
-      await viem.waitForTransactionSuccess(transferOwnershipHash)
-    case owner.address:
-      const ownerIsRootController = await root.read.controllers([owner.address])
-      if (!ownerIsRootController) {
-        const setControllerHash = await root.write.setController(
-          [owner.address, true],
-          { account: owner.account },
-        )
-        console.log(
-          `Setting final owner as controller on root contract (tx: ${setControllerHash})...`,
-        )
-        await viem.waitForTransactionSuccess(setControllerHash)
-      }
-      break
-    default:
-      console.log(
-        `WARNING: Root is owned by ${rootOwner}; cannot transfer to owner account`,
-      )
-  }
-
-  return true
-}
-
-func.id = 'Root:setup v1.0.0'
-func.tags = ['category:root', 'Root', 'Root:setup']
-func.dependencies = ['Root:contract']
-
-export default func
+  },
+  {
+    id: 'Root:setup v1.0.0',
+    tags: ['category:root', 'Root', 'Root:setup'],
+    dependencies: ['Root:contract'],
+  },
+)
