@@ -1,25 +1,30 @@
-import type { DeployFunction } from 'hardhat-deploy/types.js'
+import { artifacts, execute } from '@rocketh'
 
-const func: DeployFunction = async function (hre) {
-  const { network, viem } = hre
+export default execute(
+  async ({ deploy, namedAccounts, network }) => {
+    const { deployer } = namedAccounts
 
-  let metadataHost =
-    process.env.METADATA_HOST || 'ens-metadata-service.appspot.com'
+    let metadataHost =
+      process.env.METADATA_HOST || 'ens-metadata-service.appspot.com'
 
-  if (network.name === 'localhost') {
-    metadataHost = 'http://localhost:8080'
-  }
+    if (network.name === 'localhost') {
+      metadataHost = 'http://localhost:8080'
+    }
 
-  const metadataUrl = `${metadataHost}/name/0x{id}`
+    const metadataUrl = `${metadataHost}/name/0x{id}`
 
-  await viem.deploy('StaticMetadataService', [metadataUrl])
+    await deploy('StaticMetadataService', {
+      account: deployer,
+      artifact: artifacts.StaticMetadataService,
+      args: [metadataUrl],
+    })
 
-  return true
-}
-
-func.id = 'StaticMetadataService v1.0.0'
-func.tags = ['category:wrapper', 'StaticMetadataService']
-// technically not a dep, but we want to make sure it's deployed first for the consistent address
-func.dependencies = ['BaseRegistrarImplementation']
-
-export default func
+    return true
+  },
+  {
+    id: 'StaticMetadataService v1.0.0',
+    tags: ['category:wrapper', 'StaticMetadataService'],
+    // technically not a dep, but we want to make sure it's deployed first for the consistent address
+    dependencies: ['BaseRegistrarImplementation'],
+  },
+)
