@@ -1,29 +1,56 @@
 import { createAnvil } from '@viem/anvil'
-import { execSync } from 'child_process'
+import { executeDeployScripts, resolveConfig } from 'rocketh'
+import { createWalletClient, http } from 'viem'
 
-const server = createAnvil({
-  host: '127.0.0.1',
-  port: 8545,
+const t0 = Date.now()
+
+const anvil = createAnvil()
+await anvil.start()
+
+const hostPort = `http://${anvil.host}:${anvil.port}`
+
+const pollingInterval = 1
+
+const client = createWalletClient({
+  transport: http(hostPort),
+  pollingInterval,
 })
 
-await server.start()
+const [deployer, owner] = await client.requestAddresses()
+const accounts = { deployer, owner }
 
-const exitHandler = async (c: number) => {
-  if (process.env.CI) process.exit(c)
-  else await server.stop()
-}
+process.env.BATCH_GATEWAY_URLS = '["x-batch-gateway:true"]'
 
-process.on('exit', exitHandler)
+const env = await executeDeployScripts(
+  resolveConfig({
+    network: {
+      name: 'local',
+      tags: ['test', 'legacy', 'use_root', 'allow_unsafe'],
+      nodeUrl: hostPort,
+      fork: false,
+      pollingInterval: Math.max(1, pollingInterval) / 1000, // can't be 0
+    },
+    accounts,
+    askBeforeProceeding: false,
+    saveDeployments: false,
+    logLevel: 1,
+  }),
+)
 
-process.on('beforeExit', exitHandler)
+console.table(
+  Object.entries(env.deployments).map(([name, { address }]) => ({
+    name,
+    address,
+  })),
+)
 
-execSync('bun run hardhat --network localhost deploy', {
-  stdio: 'inherit',
-  env: {
-    ...process.env,
-    NODE_OPTIONS: '--experimental-loader ts-node/esm/transpile-only',
-    BATCH_GATEWAY_URLS: '["https://example.com/"]',
-  },
-})
+console.log(`\nReady <${Date.now() - t0}ms>`)
 
-await exitHandler(0)
+// the execa logic is completely broken and makes no sense
+// await anvil.stop();
+
+// anyway, this was launched as a child process
+// so we can just exit
+process.exit()
+
+// TODO: maybe this should be `bun run devnet`?
