@@ -1,5 +1,5 @@
 import hre from 'hardhat'
-import type { ArtifactsMap, CompilerInput } from 'hardhat/types/artifacts.js'
+import type { ArtifactMap } from 'hardhat/types/artifacts'
 import {
   bytesToHex,
   hexToBytes,
@@ -7,79 +7,6 @@ import {
   type Abi,
   type AbiFunction,
 } from 'viem'
-
-/**
- * @description Matches a function signature string to an exact ABI function.
- *
- * - Required to ensure that the ABI function is an **exact** match for the string, avoiding any potential mismatches.
- *
- * @param {Object} params
- * @param {Abi} params.artifactAbi - The ABI of the interface artifact
- * @param {string} params.fnString - The function signature string to match
- * @returns
- */
-const matchStringFunctionToAbi = ({
-  artifactAbi,
-  fnString,
-}: {
-  artifactAbi: Abi
-  fnString: string
-}) => {
-  // Extract the function name from the function signature string
-  const name = fnString.match(/(?<=function ).*?(?=\()/)![0]
-
-  // Find all functions with the same name
-  let matchingFunctions = artifactAbi.filter(
-    (abi): abi is AbiFunction => abi.type === 'function' && abi.name === name,
-  )
-  // If there is only one function with the same name, return it
-  if (matchingFunctions.length === 1) return matchingFunctions[0]
-
-  // Extract the input types as strings from the function signature string
-  const inputStrings = fnString
-    .match(/(?<=\().*?(?=\))/)![0]
-    .split(',')
-    .map((x) => x.trim())
-
-  // Filter out functions with a different number of inputs
-  matchingFunctions = matchingFunctions.filter(
-    (abi) => abi.inputs.length === inputStrings.length,
-  )
-  // If there is only one function with the same number of inputs, return it
-  if (matchingFunctions.length === 1) return matchingFunctions[0]
-
-  // Parse the input strings into input type/name
-  const parsedInputs = inputStrings.map((x) => {
-    const [type, name] = x.split(' ')
-    return { type, name }
-  })
-
-  // Filter out functions with different input types
-  matchingFunctions = matchingFunctions.filter((abi) => {
-    for (let i = 0; i < abi.inputs.length; i++) {
-      const current = parsedInputs[i]
-      const reference = abi.inputs[i]
-      // Standard match for most cases (e.g. 'uint256' === 'uint256')
-      if (reference.type === current.type) continue
-      if ('internalType' in reference && reference.internalType) {
-        // Internal types that are equal
-        if (reference.internalType === current.type) continue
-        // Internal types that are effectively equal (e.g. 'contract INameWrapperUpgrade' === 'INameWrapperUpgrade')
-        // Multiple internal type aliases can't exist in the same contract, so this is safe
-        const internalTypeName = reference.internalType.split(' ')[1]
-        if (internalTypeName === current.type) continue
-      }
-      // Not matching
-      return false
-    }
-    // 0 length input - matched by default since the filter for input length already passed
-    return true
-  })
-  // If there is only one function with the same inputs, return it
-  if (matchingFunctions.length === 1) return matchingFunctions[0]
-
-  throw new Error(`Could not find matching function for ${fnString}`)
-}
 
 /**
  * @description Gets the interface ABI that would be used in Solidity
@@ -90,59 +17,14 @@ const matchStringFunctionToAbi = ({
  * @returns The explicitly defined ABI for the interface
  */
 const getSolidityReferenceInterfaceAbi = async (
-  interfaceName: keyof ArtifactsMap,
+  interfaceName: keyof ArtifactMap,
 ) => {
-  const artifact = await hre.artifacts.readArtifact(interfaceName)
-  const fullyQualifiedNames = await hre.artifacts.getAllFullyQualifiedNames()
+  const artifact = await hre.artifacts.readArtifact(interfaceName as string)
 
-  const fullyQualifiedInterfaceName = fullyQualifiedNames.find((n) =>
-    n.endsWith(interfaceName),
-  )
-
-  if (!fullyQualifiedInterfaceName)
-    throw new Error("Couldn't find fully qualified interface name")
-
-  const buildInfo = await hre.artifacts.getBuildInfo(
-    fullyQualifiedInterfaceName,
-  )
-
-  if (!buildInfo) throw new Error("Couldn't find build info for interface")
-
-  const path = fullyQualifiedInterfaceName.split(':')[0]
-  const buildMetadata = JSON.parse(
-    (buildInfo.output.contracts[path][interfaceName] as any).metadata,
-  ) as CompilerInput
-  const { content } = buildMetadata.sources[path]
-
-  return (
-    content
-      // Remove comments - single and multi-line
-      .replace(/\/\*[\s\S]*?\*\/|\/\/.*$/gm, '')
-      // Remove structs and enums
-      .replaceAll(/((enum)|(struct)) \w+ {[^{]*?}/g, '')
-      // Match only the interface block + nested curly braces
-      .match(`interface ${interfaceName} .*?{(?:\{??[^{]*?})+`)![0]
-      // Remove the interface keyword and the interface name
-      .replace(/.*{/s, '')
-      // Remove the closing curly brace
-      .replace(/}$/s, '')
-      // Match array of all function signatures
-      .match(/function .*?;/gs)!
-      // Remove newlines and trailing semicolons
-      .map((fn) =>
-        fn
-          .split('\n')
-          .map((l) => l.trim())
-          .join('')
-          .replace(/;$/, ''),
-      )
-      // Match the function signature string to the exact ABI function
-      .map((fnString) =>
-        matchStringFunctionToAbi({
-          artifactAbi: artifact.abi as Abi,
-          fnString,
-        }),
-      )
+  // For interfaces, the artifact ABI contains only the functions explicitly defined in the interface
+  // This is exactly what we need for calculating the interface ID
+  return artifact.abi.filter(
+    (item): item is AbiFunction => item.type === 'function',
   )
 }
 
@@ -161,7 +43,7 @@ export const createInterfaceId = <iface extends Abi>(iface: iface) => {
   return bytesToHex(bytesId)
 }
 
-export const getInterfaceId = async (interfaceName: keyof ArtifactsMap) => {
+export const getInterfaceId = async (interfaceName: keyof ArtifactMap) => {
   const abi = await getSolidityReferenceInterfaceAbi(interfaceName)
   return createInterfaceId(abi)
 }
