@@ -1,13 +1,17 @@
-import hre from 'hardhat'
 import { readFile } from 'node:fs/promises'
 import {
-  type Hex,
   type Abi,
+  type Account,
   type Address,
-  sliceHex,
+  type Chain,
   concat,
   getContractAddress,
+  type Hex,
+  sliceHex,
+  type Transport,
+  type WalletClient,
 } from 'viem'
+import { getTransactionCount, waitForTransactionReceipt } from 'viem/actions'
 
 type LinkReferences = Record<
   string,
@@ -28,12 +32,14 @@ type HardhatArtifact = {
   linkReferences: LinkReferences
 }
 
-export async function deployArtifact(options: {
-  file: string | URL
-  from?: Hex
-  args?: any[]
-  libs?: Record<string, Address>
-}) {
+export async function deployArtifact(
+  walletClient: WalletClient<Transport, Chain, Account>,
+  options: {
+    file: string | URL
+    args?: any[]
+    libs?: Record<string, Address>
+  },
+) {
   const artifact = JSON.parse(await readFile(options.file, 'utf8')) as
     | ForgeArtifact
     | HardhatArtifact
@@ -48,30 +54,26 @@ export async function deployArtifact(options: {
   }
   for (const ref of Object.values(linkReferences)) {
     for (const [name, places] of Object.entries(ref)) {
-      const lib = options.libs?.[name]
-      if (!lib) throw new Error(`expected library: ${name}`)
+      const address = options.libs?.[name]
+      if (!address) throw new Error(`expected library: ${name}`)
       for (const { start, length } of places) {
         bytecode = concat([
           sliceHex(bytecode, 0, start),
-          lib,
+          address,
           sliceHex(bytecode, start + length),
         ])
       }
     }
   }
-  const walletClient = options.from
-    ? await hre.viem.getWalletClient(options.from)
-    : await hre.viem.getWalletClients().then((x) => x[0])
-  const publicClient = await hre.viem.getPublicClient()
   const nonce = BigInt(
-    await publicClient.getTransactionCount(walletClient.account),
+    await getTransactionCount(walletClient, walletClient.account),
   )
   const hash = await walletClient.deployContract({
     abi: artifact.abi,
     bytecode,
     args: options.args,
   })
-  await publicClient.waitForTransactionReceipt({ hash })
+  await waitForTransactionReceipt(walletClient, { hash })
   return getContractAddress({
     from: walletClient.account.address,
     nonce,
