@@ -2,8 +2,8 @@ import hre from 'hardhat'
 
 export function isHardhatFork() {
   return (
-    hre.network.name === 'hardhat' &&
-    'forking' in hre.network.config &&
-    !!hre.network.config.forking?.enabled
+    (hre.network as any).name === 'hardhat' &&
+    'forking' in (hre.network as any).config &&
+    !!(hre.network as any).config.forking?.enabled
   )
 }
