@@ -1,11 +1,11 @@
-import hre from 'hardhat'
 import type { Abi, AbiFunction } from 'abitype'
-import type { Artifact, ArtifactsMap } from 'hardhat/types/index.js'
-import { loadFixture } from '@nomicfoundation/hardhat-toolbox-viem/network-helpers.js'
-import { expect } from 'chai'
+import hre from 'hardhat'
+import type { Artifact, ArtifactMap } from 'hardhat/types/artifacts'
 
-export function runSolidityTests<N extends keyof ArtifactsMap>(name: N) {
-  const artifact: Artifact = hre.artifacts.readArtifactSync(name)
+const connection = await hre.network.connect()
+
+export async function runSolidityTests<N extends keyof ArtifactMap>(name: N) {
+  const artifact: Artifact = await hre.artifacts.readArtifact(name)
   const abi: Abi = artifact.abi
   const tests = abi.filter(
     (x): x is AbiFunction => x.type === 'function' && x.name.startsWith('test'),
@@ -13,15 +13,15 @@ export function runSolidityTests<N extends keyof ArtifactsMap>(name: N) {
   if (!tests.length) throw new Error(`no tests: ${name}`)
 
   async function fixture() {
-    const publicClient = await hre.viem.getPublicClient()
-    const contract = await hre.viem.deployContract(artifact.contractName)
+    const publicClient = await connection.viem.getPublicClient()
+    const contract = await connection.viem.deployContract(name)
     return { publicClient, contract }
   }
 
   describe(name, () => {
     tests.forEach((fn) => {
       it(fn.name, async () => {
-        const F = await loadFixture(fixture)
+        const F = await connection.networkHelpers.loadFixture(fixture)
         if (fn.name.startsWith('testFail')) {
           await expect(
             F.publicClient.readContract({
