@@ -1,33 +1,37 @@
 import hre from 'hardhat'
-import { loadFixture } from '@nomicfoundation/hardhat-toolbox-viem/network-helpers.js'
-import { expect } from 'chai'
+
 import { dnsEncodeName } from '../fixtures/dnsEncodeName.js'
-import { serveBatchGateway } from '../fixtures/localBatchGateway.js'
 import { shortCoin } from '../fixtures/ensip19.js'
 import { isHardhatFork } from '../fixtures/forked.js'
+import { serveBatchGateway } from '../fixtures/localBatchGateway.js'
 import { ENS_REGISTRY, KNOWN_PRIMARIES, KNOWN_RESOLUTIONS } from './mainnet.js'
 import { bundleCalls, makeResolutions } from '../utils/resolutions.js'
 
 // $ bun run test:remote
 
+const connection = await hre.network.connect()
+
 async function fixture() {
   const bg = await serveBatchGateway()
-  after(bg.shutdown)
-  const [owner] = await hre.viem.getWalletClients()
-  const batchGatewayProvider = await hre.viem.deployContract(
+  afterAll(bg.shutdown)
+
+  const [owner] = await connection.viem.getWalletClients()
+  const batchGatewayProvider = await connection.viem.deployContract(
     'GatewayProvider',
     [owner.account.address, [bg.localBatchGatewayUrl]],
   )
-  return hre.viem.deployContract(
+
+  return connection.viem.deployContract(
     'UniversalResolver',
     [owner.account.address, ENS_REGISTRY, batchGatewayProvider.address],
     {
       client: {
-        public: await hre.viem.getPublicClient({ ccipRead: undefined }),
+        public: await connection.viem.getPublicClient({ ccipRead: undefined }),
       },
     },
   )
 }
+const loadFixture = async () => connection.networkHelpers.loadFixture(fixture)
 
 ;(isHardhatFork() ? describe : describe.skip)(
   'UniversalResolver @ mainnet',
@@ -37,7 +41,7 @@ async function fixture() {
         const calls = makeResolutions(x)
         it(`${x.title}: ${x.name} [${calls.length}]`, async () => {
           const bundle = bundleCalls(calls)
-          const F = await loadFixture(fixture)
+          const F = await loadFixture()
           const [answer] = await F.read.resolve([
             dnsEncodeName(x.name),
             bundle.call,
@@ -49,7 +53,7 @@ async function fixture() {
     describe('reverse()', () => {
       for (const x of KNOWN_PRIMARIES) {
         it(`${x.title}: ${shortCoin(x.coinType)} ${x.address}`, async () => {
-          const F = await loadFixture(fixture)
+          const F = await loadFixture()
           const promise = F.read.reverse([x.address, x.coinType])
           if (x.expectError) {
             await expect(promise).rejects.toThrow()
