@@ -1,4 +1,3 @@
-import { expect } from 'chai'
 import {
   type Hex,
   decodeFunctionResult,
@@ -7,6 +6,7 @@ import {
   getAddress,
   namehash,
   parseAbi,
+  toHex,
 } from 'viem'
 import { COIN_TYPE_ETH, shortCoin } from '../fixtures/ensip19.js'
 
@@ -107,8 +107,16 @@ export function bundleCalls(resolutions: KnownResolution[]): KnownBundle {
     }),
     answer: encodeFunctionResult({
       abi: RESOLVE_MULTICALL,
-      // TODO: fix when we can use newer viem version
-      result: [resolutions.map((x) => x.answer)] as never,
+      functionName: 'multicall',
+      result: resolutions.map((x) => {
+        let answer = x.answer
+        if (typeof answer === 'string') {
+          return answer
+        } else {
+          answer = toHex(answer)
+        }
+        return answer
+      }),
     }),
     resolutions,
     unbundle: (data) =>
@@ -148,7 +156,7 @@ export function makeResolutions(p: KnownProfile): KnownResolution[] {
             abi,
             functionName,
             // TODO: fix when we can use newer viem version
-            result: [value] as never,
+            result: value,
           }),
           expect(data) {
             const actual = decodeFunctionResult({
@@ -178,7 +186,7 @@ export function makeResolutions(p: KnownProfile): KnownResolution[] {
             abi,
             functionName,
             // TODO: fix when we can use newer viem version
-            result: [value] as never,
+            result: value,
           }),
           expect(data) {
             const actual = decodeFunctionResult({
@@ -213,9 +221,13 @@ export function makeResolutions(p: KnownProfile): KnownResolution[] {
           abi,
           functionName,
           // TODO: fix when we can use newer viem version
-          result: [value] as never,
+          result: value,
         }),
         expect(data) {
+          // Handle empty response case
+          if (data === '0x' && value === undefined) {
+            return // Empty response is expected
+          }
           const actual = decodeFunctionResult({
             abi,
             functionName,
@@ -247,14 +259,22 @@ export function makeResolutions(p: KnownProfile): KnownResolution[] {
         abi,
         functionName,
         // TODO: fix when we can use newer viem version
-        result: [value] as never,
+        result: value,
       }),
       expect(data) {
+        // Handle empty response case
+        if (data === '0x' && value === undefined) {
+          return // Empty response is expected
+        }
         const actual = decodeFunctionResult({
           abi,
           functionName,
           data,
         })
+        // Handle case where empty string is returned but undefined was expected
+        if (actual === '' && value === undefined) {
+          return // Empty string decoded as undefined
+        }
         expect(actual, this.desc).toStrictEqual(value)
       },
       write: encodeFunctionData({
