@@ -1,9 +1,10 @@
-import { loadFixture } from '@nomicfoundation/hardhat-toolbox-viem/network-helpers.js'
-import { expect } from 'chai'
+import type { Fixture } from '@nomicfoundation/hardhat-network-helpers/types'
+import type { NetworkConnection } from 'hardhat/types/network'
 import { getAddress, labelhash, namehash, zeroAddress } from 'viem'
+
 import { DAY, FUSES } from '../fixtures/constants.js'
 import { dnsEncodeName } from '../fixtures/dnsEncodeName.js'
-import { toTokenId } from '../fixtures/utils.js'
+import { getAccounts, toTokenId } from '../fixtures/utils.js'
 import { deployNameWrapperFixture } from './fixtures/deploy.js'
 
 // States
@@ -19,152 +20,649 @@ import { deployNameWrapperFixture } from './fixtures/deploy.js'
 // 0001 = Parent burned parent's CU (PCU)
 // Each can be combined to represent multiple states
 
-const {
-  CANNOT_UNWRAP,
-  CANNOT_SET_RESOLVER,
-  PARENT_CANNOT_CONTROL,
-  CAN_DO_EVERYTHING,
-  IS_DOT_ETH,
-} = FUSES
+export const shouldRespectConstraints = (connection: NetworkConnection) => {
+  const {
+    CANNOT_UNWRAP,
+    CANNOT_SET_RESOLVER,
+    PARENT_CANNOT_CONTROL,
+    CAN_DO_EVERYTHING,
+    IS_DOT_ETH,
+  } = FUSES
 
-const GRACE_PERIOD = 90n * DAY
-const MAX_EXPIRY = 2n ** 64n - 1n
+  const GRACE_PERIOD = 90n * DAY
+  const MAX_EXPIRY = 2n ** 64n - 1n
 
-const parentLabel = 'test1'
-const parentLabelHash = labelhash(parentLabel)
-const parentLabelId = toTokenId(parentLabelHash)
-const parentNode = namehash('test1.eth')
-const parentNodeId = toTokenId(parentNode)
-const childNode = namehash('sub.test1.eth')
-const childNodeId = toTokenId(childNode)
-const childLabel = 'sub'
-const childLabelHash = labelhash(childLabel)
+  const parentLabel = 'test1'
+  const parentLabelHash = labelhash(parentLabel)
+  const parentLabelId = toTokenId(parentLabelHash)
+  const parentNode = namehash('test1.eth')
+  const parentNodeId = toTokenId(parentNode)
+  const childNode = namehash('sub.test1.eth')
+  const childNodeId = toTokenId(childNode)
+  const childLabel = 'sub'
+  const childLabelHash = labelhash(childLabel)
 
-async function baseFixture() {
-  const initial = await loadFixture(deployNameWrapperFixture)
+  async function baseFixture() {
+    const initial = await deployNameWrapperFixture(connection)
+    const accounts = await getAccounts(connection)
+    const publicClient = await connection.viem.getPublicClient()
+    const testClient = await connection.viem.getTestClient()
 
-  await initial.baseRegistrar.write.setApprovalForAll([
-    initial.nameWrapper.address,
-    true,
-  ])
-
-  return initial
-}
-
-// Reusable state setup
-const setupState = ({
-  parentFuses,
-  childFuses,
-  childExpiry,
-}: {
-  parentFuses: number
-  childFuses: number
-  childExpiry: bigint
-}) =>
-  async function setupStateFixture() {
-    const initial = await loadFixture(baseFixture)
-    const { baseRegistrar, nameWrapper, accounts } = initial
-
-    await baseRegistrar.write.register([
-      parentLabelId,
-      accounts[0].address,
-      DAY,
-    ])
-    await nameWrapper.write.wrapETH2LD([
-      parentLabel,
-      accounts[0].address,
-      parentFuses,
-      zeroAddress,
+    await initial.baseRegistrar.write.setApprovalForAll([
+      initial.nameWrapper.address,
+      true,
     ])
 
-    await nameWrapper.write.setSubnodeOwner([
-      parentNode,
-      childLabel,
-      accounts[1].address,
-      childFuses,
-      childExpiry, // Expired ??
-    ])
-
-    return initial
+    return { ...initial, accounts, publicClient, testClient }
   }
 
-// Reusable state setup
-const setupStateUnexpired = ({
-  parentFuses,
-  childFuses,
-}: {
-  parentFuses: number
-  childFuses: number
-}) =>
-  async function setupStateUnexpiredFixture() {
-    const initial = await loadFixture(baseFixture)
-    const { baseRegistrar, nameWrapper, accounts } = initial
+  const loadFixture = async <T>(fixture: Fixture<T>): Promise<T> =>
+    connection.networkHelpers.loadFixture(fixture)
 
-    await baseRegistrar.write.register([
-      parentLabelId,
-      accounts[0].address,
-      DAY * 2n,
-    ])
-    const parentExpiry = await baseRegistrar.read.nameExpires([parentLabelId])
-    await nameWrapper.write.wrapETH2LD([
-      parentLabel,
-      accounts[0].address,
-      parentFuses,
-      zeroAddress,
-    ])
+  // Reusable state setup
+  const setupState = ({
+    parentFuses,
+    childFuses,
+    childExpiry,
+  }: {
+    parentFuses: number
+    childFuses: number
+    childExpiry: bigint
+  }) =>
+    async function setupStateFixture() {
+      const initial = await baseFixture()
+      const { baseRegistrar, nameWrapper, accounts } = initial
 
-    await nameWrapper.write.setSubnodeOwner([
-      parentNode,
-      childLabel,
-      accounts[1].address,
-      childFuses,
-      parentExpiry - DAY, // Expires a day before parent
-    ])
+      await baseRegistrar.write.register([
+        parentLabelId,
+        accounts[0].address,
+        DAY,
+      ])
+      await nameWrapper.write.wrapETH2LD([
+        parentLabel,
+        accounts[0].address,
+        parentFuses,
+        zeroAddress,
+      ])
 
-    return initial
+      await nameWrapper.write.setSubnodeOwner([
+        parentNode,
+        childLabel,
+        accounts[1].address,
+        childFuses,
+        childExpiry, // Expired ??
+      ])
+
+      return initial
+    }
+
+  // Reusable state setup
+  const setupStateUnexpired = ({
+    parentFuses,
+    childFuses,
+  }: {
+    parentFuses: number
+    childFuses: number
+  }) =>
+    async function setupStateUnexpiredFixture() {
+      const initial = await baseFixture()
+      const { baseRegistrar, nameWrapper, accounts } = initial
+
+      await baseRegistrar.write.register([
+        parentLabelId,
+        accounts[0].address,
+        DAY * 2n,
+      ])
+      const parentExpiry = await baseRegistrar.read.nameExpires([parentLabelId])
+      await nameWrapper.write.wrapETH2LD([
+        parentLabel,
+        accounts[0].address,
+        parentFuses,
+        zeroAddress,
+      ])
+
+      await nameWrapper.write.setSubnodeOwner([
+        parentNode,
+        childLabel,
+        accounts[1].address,
+        childFuses,
+        parentExpiry - DAY, // Expires a day before parent
+      ])
+
+      return initial
+    }
+
+  // Expired, nothing burnt.
+  const setupState0000DW = setupState({
+    parentFuses: CAN_DO_EVERYTHING,
+    childFuses: CAN_DO_EVERYTHING,
+    childExpiry: 0n,
+  })
+  const setupState0001PCU = setupState({
+    parentFuses: CANNOT_UNWRAP,
+    childFuses: CAN_DO_EVERYTHING,
+    childExpiry: 0n,
+  })
+  const setupState1000NE = setupStateUnexpired({
+    childFuses: CAN_DO_EVERYTHING,
+    parentFuses: CAN_DO_EVERYTHING,
+  })
+  const setupState1001NE_PCU = setupStateUnexpired({
+    childFuses: CAN_DO_EVERYTHING,
+    parentFuses: CANNOT_UNWRAP,
+  })
+  const setupState1011NE_PCC_PCU = setupStateUnexpired({
+    childFuses: PARENT_CANNOT_CONTROL,
+    parentFuses: CANNOT_UNWRAP,
+  })
+  const setupState1111NE_CU_PCC_PCU = setupStateUnexpired({
+    childFuses: PARENT_CANNOT_CONTROL | CANNOT_UNWRAP,
+    parentFuses: CANNOT_UNWRAP,
+  })
+
+  type BaseTestParameters = {
+    fixture: () => ReturnType<typeof baseFixture>
   }
 
-// Expired, nothing burnt.
-const setupState0000DW = setupState({
-  parentFuses: CAN_DO_EVERYTHING,
-  childFuses: CAN_DO_EVERYTHING,
-  childExpiry: 0n,
-})
-const setupState0001PCU = setupState({
-  parentFuses: CANNOT_UNWRAP,
-  childFuses: CAN_DO_EVERYTHING,
-  childExpiry: 0n,
-})
-const setupState1000NE = setupStateUnexpired({
-  childFuses: CAN_DO_EVERYTHING,
-  parentFuses: CAN_DO_EVERYTHING,
-})
-const setupState1001NE_PCU = setupStateUnexpired({
-  childFuses: CAN_DO_EVERYTHING,
-  parentFuses: CANNOT_UNWRAP,
-})
-const setupState1011NE_PCC_PCU = setupStateUnexpired({
-  childFuses: PARENT_CANNOT_CONTROL,
-  parentFuses: CANNOT_UNWRAP,
-})
-const setupState1111NE_CU_PCC_PCU = setupStateUnexpired({
-  childFuses: PARENT_CANNOT_CONTROL | CANNOT_UNWRAP,
-  parentFuses: CANNOT_UNWRAP,
-})
+  // Reusable tests
+  const parentCanExtend = ({
+    fixture,
+    isNotExpired,
+  }: BaseTestParameters & {
+    isNotExpired?: boolean
+  }) => {
+    if (isNotExpired) {
+      it('Child should have an expiry < parent', async () => {
+        const { nameWrapper, baseRegistrar, publicClient } = await loadFixture(
+          fixture,
+        )
 
-type BaseTestParameters = {
-  fixture: () => ReturnType<typeof baseFixture>
-}
+        const [, , childExpiry] = await nameWrapper.read.getData([childNodeId])
+        const parentExpiry = await baseRegistrar.read.nameExpires([
+          parentLabelId,
+        ])
+        expect(childExpiry).toBeLessThan(parentExpiry)
 
-// Reusable tests
-const parentCanExtend = ({
-  fixture,
-  isNotExpired,
-}: BaseTestParameters & {
-  isNotExpired?: boolean
-}) => {
-  if (isNotExpired) {
-    it('Child should have an expiry < parent', async () => {
+        const timestamp = await publicClient.getBlock().then((b) => b.timestamp)
+        expect(childExpiry).toBeGreaterThan(timestamp)
+      })
+    } else {
+      it('Child should have a 0 expiry before extending', async () => {
+        const { nameWrapper } = await loadFixture(fixture)
+
+        const [, , expiryBefore] = await nameWrapper.read.getData([childNodeId])
+        expect(expiryBefore).toEqual(0n)
+      })
+    }
+
+    it('Parent can extend expiry with setChildFuses()', async () => {
+      const { nameWrapper, baseRegistrar } = await loadFixture(fixture)
+
+      const parentExpiry = await baseRegistrar.read.nameExpires([parentLabelId])
+
+      await nameWrapper.write.setChildFuses([
+        parentNode,
+        childLabelHash,
+        CAN_DO_EVERYTHING,
+        MAX_EXPIRY,
+      ])
+
+      const [, , expiry] = await nameWrapper.read.getData([childNodeId])
+
+      expect(expiry).toEqual(parentExpiry + GRACE_PERIOD)
+    })
+
+    it('Parent can extend expiry with setSubnodeOwner()', async () => {
+      const { nameWrapper, baseRegistrar, accounts } = await loadFixture(
+        fixture,
+      )
+
+      const parentExpiry = await baseRegistrar.read.nameExpires([parentLabelId])
+
+      await nameWrapper.write.setSubnodeOwner([
+        parentNode,
+        childLabel,
+        accounts[1].address,
+        CAN_DO_EVERYTHING,
+        MAX_EXPIRY,
+      ])
+
+      const [, , expiry] = await nameWrapper.read.getData([childNodeId])
+
+      expect(expiry).toEqual(parentExpiry + GRACE_PERIOD)
+    })
+
+    it('Parent can extend expiry with setSubnodeRecord()', async () => {
+      const { nameWrapper, baseRegistrar, accounts } = await loadFixture(
+        fixture,
+      )
+
+      const parentExpiry = await baseRegistrar.read.nameExpires([parentLabelId])
+
+      await nameWrapper.write.setSubnodeRecord([
+        parentNode,
+        childLabel,
+        accounts[1].address,
+        zeroAddress,
+        0n,
+        CAN_DO_EVERYTHING,
+        MAX_EXPIRY,
+      ])
+
+      const [, , expiry] = await nameWrapper.read.getData([childNodeId])
+
+      expect(expiry).toEqual(parentExpiry + GRACE_PERIOD)
+    })
+  }
+
+  const parentCannotBurnFusesOrPCC = ({ fixture }: BaseTestParameters) => {
+    it('Parent cannot burn fuses with setChildFuses()', async () => {
+      const { nameWrapper } = await loadFixture(fixture)
+
+      await expect(
+        nameWrapper.write.setChildFuses([
+          parentNode,
+          childLabelHash,
+          CANNOT_UNWRAP | PARENT_CANNOT_CONTROL,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
+    })
+
+    it('Parent cannot burn fuses with setSubnodeOwner()', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
+
+      await expect(
+        nameWrapper.write.setSubnodeOwner([
+          parentNode,
+          childLabel,
+          accounts[1].address,
+          CANNOT_UNWRAP | PARENT_CANNOT_CONTROL,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
+    })
+
+    it('Parent cannot burn fuses with setSubnodeRecord()', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
+
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
+          parentNode,
+          childLabel,
+          accounts[1].address,
+          zeroAddress,
+          0n,
+          CANNOT_UNWRAP | PARENT_CANNOT_CONTROL,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
+    })
+  }
+
+  const parentCanReplaceOwner = ({ fixture }: BaseTestParameters) => {
+    it('Parent can replace owner with setSubnodeOwner()', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
+
+      await expect(
+        nameWrapper.read.ownerOf([childNodeId]),
+      ).resolves.toEqualAddress(accounts[1].address)
+
+      await nameWrapper.write.setSubnodeOwner([
+        parentNode,
+        childLabel,
+        accounts[0].address,
+        CAN_DO_EVERYTHING,
+        0n,
+      ])
+
+      await expect(
+        nameWrapper.read.ownerOf([childNodeId]),
+      ).resolves.toEqualAddress(accounts[0].address)
+    })
+
+    it('Parent can replace owner with setSubnodeRecord()', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
+
+      await expect(
+        nameWrapper.read.ownerOf([childNodeId]),
+      ).resolves.toEqualAddress(accounts[1].address)
+
+      await nameWrapper.write.setSubnodeRecord([
+        parentNode,
+        childLabel,
+        accounts[0].address,
+        zeroAddress,
+        0n,
+        CAN_DO_EVERYTHING,
+        0n,
+      ])
+
+      await expect(
+        nameWrapper.read.ownerOf([childNodeId]),
+      ).resolves.toEqualAddress(accounts[0].address)
+    })
+  }
+
+  const parentCanUnwrapChild = ({ fixture }: BaseTestParameters) => {
+    it('Parent can unwrap owner with setSubnodeRecord() and then unwrap', async () => {
+      const { ensRegistry, nameWrapper, accounts } = await loadFixture(fixture)
+
+      //check previous owners
+      await expect(
+        nameWrapper.read.ownerOf([childNodeId]),
+      ).resolves.toEqualAddress(accounts[1].address)
+      await expect(ensRegistry.read.owner([childNode])).resolves.toEqualAddress(
+        nameWrapper.address,
+      )
+
+      await nameWrapper.write.setSubnodeRecord([
+        parentNode,
+        childLabel,
+        accounts[0].address,
+        zeroAddress,
+        0n,
+        CAN_DO_EVERYTHING,
+        0n,
+      ])
+
+      await nameWrapper.write.unwrap([
+        parentNode,
+        childLabelHash,
+        accounts[0].address,
+      ])
+
+      await expect(
+        nameWrapper.read.ownerOf([childNodeId]),
+      ).resolves.toEqualAddress(zeroAddress)
+      await expect(ensRegistry.read.owner([childNode])).resolves.toEqualAddress(
+        accounts[0].address,
+      )
+    })
+  }
+
+  const parentCannotBurnParentControlledFuses = ({
+    fixture,
+  }: BaseTestParameters) => {
+    it('Parent cannot burn parent-controlled fuses', async () => {
+      const { nameWrapper } = await loadFixture(fixture)
+
+      await expect(
+        nameWrapper.write.setChildFuses([
+          parentNode,
+          childLabelHash,
+          1 << 18,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
+    })
+  }
+
+  const ownerIsOwnerWhenExpired = ({ fixture }: BaseTestParameters) => {
+    it('Owner is still owner when expired', async () => {
+      const { nameWrapper, accounts, publicClient } = await loadFixture(fixture)
+
+      const timestamp = await publicClient.getBlock().then((b) => b.timestamp)
+      const [, , expiry] = await nameWrapper.read.getData([childNodeId])
+
+      expect(expiry).toBeLessThan(timestamp)
+
+      await expect(
+        nameWrapper.read.ownerOf([childNodeId]),
+      ).resolves.toEqualAddress(accounts[1].address)
+    })
+  }
+
+  const ownerCannotBurnFuses = ({ fixture }: BaseTestParameters) => {
+    it('Owner cannot burn CU because PCC is not burned', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
+
+      await expect(
+        nameWrapper.write.setFuses([childNode, CANNOT_UNWRAP], {
+          account: accounts[1],
+        }),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
+    })
+
+    it('Owner cannot burn other fuses because CU and PCC are not burned', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
+
+      await expect(
+        nameWrapper.write.setFuses([childNode, CANNOT_SET_RESOLVER], {
+          account: accounts[1],
+        }),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
+    })
+  }
+
+  const ownerCanUnwrap = ({ fixture }: BaseTestParameters) => {
+    it('Owner can unwrap', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
+
+      await nameWrapper.write.unwrap(
+        [parentNode, childLabelHash, accounts[1].address],
+        { account: accounts[1] },
+      )
+
+      await expect(
+        nameWrapper.read.ownerOf([childNodeId]),
+      ).resolves.toEqualAddress(zeroAddress)
+    })
+  }
+
+  const parentCanBurnParentControlledFusesWithExpiry = ({
+    fixture,
+  }: BaseTestParameters) => {
+    it('Parent cannot burn parent-controlled fuses as they reset to 0', async () => {
+      const { nameWrapper } = await loadFixture(fixture)
+
+      await nameWrapper.write.setChildFuses([
+        parentNode,
+        childLabelHash,
+        1 << 18,
+        0n,
+      ])
+
+      // expired names get normalised to 0
+      const [, fuses] = await nameWrapper.read.getData([childNodeId])
+      expect(fuses).toEqual(0)
+    })
+
+    it('Parent can burn parent-controlled fuses, if expiry is extended', async () => {
+      const { nameWrapper } = await loadFixture(fixture)
+
+      await nameWrapper.write.setChildFuses([
+        parentNode,
+        childLabelHash,
+        1 << 18,
+        MAX_EXPIRY,
+      ])
+
+      const [, fuses] = await nameWrapper.read.getData([childNodeId])
+      expect(fuses).toEqual(1 << 18)
+    })
+  }
+
+  const parentCanBurnFusesOrPCC = ({ fixture }: BaseTestParameters) => {
+    it('Parent can burn fuses with setChildFuses()', async () => {
+      const { nameWrapper } = await loadFixture(fixture)
+
+      await nameWrapper.write.setChildFuses([
+        parentNode,
+        childLabelHash,
+        CANNOT_UNWRAP | PARENT_CANNOT_CONTROL | CANNOT_SET_RESOLVER,
+        0n,
+      ])
+
+      const [, fuses] = await nameWrapper.read.getData([childNodeId])
+      expect(fuses).toEqual(
+        CANNOT_UNWRAP | PARENT_CANNOT_CONTROL | CANNOT_SET_RESOLVER,
+      )
+    })
+
+    it('Parent can burn fuses with setSubnodeOwner()', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
+
+      await nameWrapper.write.setSubnodeOwner([
+        parentNode,
+        childLabel,
+        accounts[1].address,
+        CANNOT_UNWRAP | PARENT_CANNOT_CONTROL | CANNOT_SET_RESOLVER,
+        0n,
+      ])
+
+      const [, fuses] = await nameWrapper.read.getData([childNodeId])
+      expect(fuses).toEqual(
+        CANNOT_UNWRAP | PARENT_CANNOT_CONTROL | CANNOT_SET_RESOLVER,
+      )
+    })
+
+    it('Parent can burn fuses with setSubnodeRecord()', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
+
+      await nameWrapper.write.setSubnodeRecord([
+        parentNode,
+        childLabel,
+        accounts[1].address,
+        zeroAddress,
+        0n,
+        CANNOT_UNWRAP | PARENT_CANNOT_CONTROL | CANNOT_SET_RESOLVER,
+        0n,
+      ])
+
+      const [, fuses] = await nameWrapper.read.getData([childNodeId])
+      expect(fuses).toEqual(
+        CANNOT_UNWRAP | PARENT_CANNOT_CONTROL | CANNOT_SET_RESOLVER,
+      )
+    })
+
+    it('Parent cannot burn fuses if PCC is not burnt too', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
+
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
+          parentNode,
+          childLabel,
+          accounts[1].address,
+          zeroAddress,
+          0n,
+          CANNOT_UNWRAP | CANNOT_SET_RESOLVER,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
+
+      await expect(
+        nameWrapper.write.setChildFuses([
+          parentNode,
+          childLabelHash,
+          CANNOT_UNWRAP | CANNOT_SET_RESOLVER,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
+
+      await expect(
+        nameWrapper.write.setSubnodeOwner([
+          parentNode,
+          childLabel,
+          accounts[1].address,
+          CANNOT_UNWRAP | CANNOT_SET_RESOLVER,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
+    })
+  }
+
+  const parentCanBurnParentControlledFuses = ({
+    fixture,
+  }: BaseTestParameters) => {
+    it('Parent can burn parent-controlled fuses', async () => {
+      const { nameWrapper } = await loadFixture(fixture)
+
+      await nameWrapper.write.setChildFuses([
+        parentNode,
+        childLabelHash,
+        1 << 18,
+        0n,
+      ])
+
+      const [, fuses] = await nameWrapper.read.getData([childNodeId])
+      expect(fuses).toEqual(1 << 18)
+    })
+  }
+
+  const testStateTransition1000to1010 = ({ fixture }: BaseTestParameters) => {
+    it('1000 => 1010 - Parent cannot burn PCC with setChildFuses()', async () => {
+      const { nameWrapper } = await loadFixture(fixture)
+
+      await expect(
+        nameWrapper.write.setChildFuses([
+          parentNode,
+          childLabelHash,
+          PARENT_CANNOT_CONTROL,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
+    })
+
+    it('1000 => 1010 - Parent cannot burn PCC with setSubnodeOwner()', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
+
+      await expect(
+        nameWrapper.write.setSubnodeOwner([
+          parentNode,
+          childLabel,
+          accounts[1].address,
+          PARENT_CANNOT_CONTROL,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
+    })
+
+    it('1000 => 1010 - Parent cannot burn PCC with setSubnodeRecord()', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
+
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
+          parentNode,
+          childLabel,
+          accounts[1].address,
+          zeroAddress,
+          0n,
+          PARENT_CANNOT_CONTROL,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
+    })
+  }
+
+  const parentCanExtendWithSetChildFuses = ({
+    fixture,
+  }: BaseTestParameters) => {
+    it('Child should have a <parent expiry before extending', async () => {
       const { nameWrapper, baseRegistrar, publicClient } = await loadFixture(
         fixture,
       )
@@ -176,759 +674,317 @@ const parentCanExtend = ({
       const timestamp = await publicClient.getBlock().then((b) => b.timestamp)
       expect(childExpiry).toBeGreaterThan(timestamp)
     })
-  } else {
-    it('Child should have a 0 expiry before extending', async () => {
-      const { nameWrapper } = await loadFixture(fixture)
 
-      const [, , expiryBefore] = await nameWrapper.read.getData([childNodeId])
-      expect(expiryBefore).toEqual(0n)
+    it('Parent can extend expiry with setChildFuses()', async () => {
+      const { nameWrapper, baseRegistrar } = await loadFixture(fixture)
+
+      const parentExpiry = await baseRegistrar.read.nameExpires([parentLabelId])
+
+      await nameWrapper.write.setChildFuses([
+        parentNode,
+        childLabelHash,
+        CAN_DO_EVERYTHING,
+        MAX_EXPIRY,
+      ])
+
+      const [, , expiry] = await nameWrapper.read.getData([childNodeId])
+
+      expect(expiry).toEqual(parentExpiry + GRACE_PERIOD)
+    })
+
+    it('Parent cannot extend expiry with setSubnodeOwner()', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
+
+      await expect(
+        nameWrapper.write.setSubnodeOwner([
+          parentNode,
+          childLabel,
+          accounts[1].address,
+          CAN_DO_EVERYTHING,
+          MAX_EXPIRY,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
+    })
+
+    it('Parent cannot extend expiry with setSubnodeRecord()', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
+
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
+          parentNode,
+          childLabel,
+          accounts[1].address,
+          zeroAddress,
+          0n,
+          CAN_DO_EVERYTHING,
+          MAX_EXPIRY,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
     })
   }
 
-  it('Parent can extend expiry with setChildFuses()', async () => {
-    const { nameWrapper, baseRegistrar, accounts } = await loadFixture(fixture)
-
-    const parentExpiry = await baseRegistrar.read.nameExpires([parentLabelId])
-
-    await nameWrapper.write.setChildFuses([
-      parentNode,
-      childLabelHash,
-      CAN_DO_EVERYTHING,
-      MAX_EXPIRY,
-    ])
-
-    const [, , expiry] = await nameWrapper.read.getData([childNodeId])
-
-    expect(expiry).toEqual(parentExpiry + GRACE_PERIOD)
-  })
-
-  it('Parent can extend expiry with setSubnodeOwner()', async () => {
-    const { nameWrapper, baseRegistrar, accounts } = await loadFixture(fixture)
-
-    const parentExpiry = await baseRegistrar.read.nameExpires([parentLabelId])
-
-    await nameWrapper.write.setSubnodeOwner([
-      parentNode,
-      childLabel,
-      accounts[1].address,
-      CAN_DO_EVERYTHING,
-      MAX_EXPIRY,
-    ])
-
-    const [, , expiry] = await nameWrapper.read.getData([childNodeId])
-
-    expect(expiry).toEqual(parentExpiry + GRACE_PERIOD)
-  })
-
-  it('Parent can extend expiry with setSubnodeRecord()', async () => {
-    const { nameWrapper, baseRegistrar, accounts } = await loadFixture(fixture)
-
-    const parentExpiry = await baseRegistrar.read.nameExpires([parentLabelId])
-
-    await nameWrapper.write.setSubnodeRecord([
-      parentNode,
-      childLabel,
-      accounts[1].address,
-      zeroAddress,
-      0n,
-      CAN_DO_EVERYTHING,
-      MAX_EXPIRY,
-    ])
-
-    const [, , expiry] = await nameWrapper.read.getData([childNodeId])
-
-    expect(expiry).toEqual(parentExpiry + GRACE_PERIOD)
-  })
-}
-
-const parentCannotBurnFusesOrPCC = ({ fixture }: BaseTestParameters) => {
-  it('Parent cannot burn fuses with setChildFuses()', async () => {
-    const { nameWrapper } = await loadFixture(fixture)
-
-    await expect(nameWrapper)
-      .write('setChildFuses', [
-        parentNode,
-        childLabelHash,
-        CANNOT_UNWRAP | PARENT_CANNOT_CONTROL,
-        0n,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-  })
-
-  it('Parent cannot burn fuses with setSubnodeOwner()', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
-
-    await expect(nameWrapper)
-      .write('setSubnodeOwner', [
-        parentNode,
-        childLabel,
-        accounts[1].address,
-        CANNOT_UNWRAP | PARENT_CANNOT_CONTROL,
-        0n,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-  })
-
-  it('Parent cannot burn fuses with setSubnodeRecord()', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
-
-    await expect(nameWrapper)
-      .write('setSubnodeRecord', [
-        parentNode,
-        childLabel,
-        accounts[1].address,
-        zeroAddress,
-        0n,
-        CANNOT_UNWRAP | PARENT_CANNOT_CONTROL,
-        0n,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-  })
-}
-
-const parentCanReplaceOwner = ({ fixture }: BaseTestParameters) => {
-  it('Parent can replace owner with setSubnodeOwner()', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
-
-    await expect(
-      nameWrapper.read.ownerOf([childNodeId]),
-    ).resolves.toEqualAddress(accounts[1].address)
-
-    await nameWrapper.write.setSubnodeOwner([
-      parentNode,
-      childLabel,
-      accounts[0].address,
-      CAN_DO_EVERYTHING,
-      0n,
-    ])
-
-    await expect(
-      nameWrapper.read.ownerOf([childNodeId]),
-    ).resolves.toEqualAddress(accounts[0].address)
-  })
-
-  it('Parent can replace owner with setSubnodeRecord()', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
-
-    await expect(
-      nameWrapper.read.ownerOf([childNodeId]),
-    ).resolves.toEqualAddress(accounts[1].address)
-
-    await nameWrapper.write.setSubnodeRecord([
-      parentNode,
-      childLabel,
-      accounts[0].address,
-      zeroAddress,
-      0n,
-      CAN_DO_EVERYTHING,
-      0n,
-    ])
-
-    await expect(
-      nameWrapper.read.ownerOf([childNodeId]),
-    ).resolves.toEqualAddress(accounts[0].address)
-  })
-}
-
-const parentCanUnwrapChild = ({ fixture }: BaseTestParameters) => {
-  it('Parent can unwrap owner with setSubnodeRecord() and then unwrap', async () => {
-    const { ensRegistry, nameWrapper, accounts } = await loadFixture(fixture)
-
-    //check previous owners
-    await expect(
-      nameWrapper.read.ownerOf([childNodeId]),
-    ).resolves.toEqualAddress(accounts[1].address)
-    await expect(ensRegistry.read.owner([childNode])).resolves.toEqualAddress(
-      nameWrapper.address,
-    )
-
-    await nameWrapper.write.setSubnodeRecord([
-      parentNode,
-      childLabel,
-      accounts[0].address,
-      zeroAddress,
-      0n,
-      CAN_DO_EVERYTHING,
-      0n,
-    ])
-
-    await nameWrapper.write.unwrap([
-      parentNode,
-      childLabelHash,
-      accounts[0].address,
-    ])
-
-    await expect(
-      nameWrapper.read.ownerOf([childNodeId]),
-    ).resolves.toEqualAddress(zeroAddress)
-    await expect(ensRegistry.read.owner([childNode])).resolves.toEqualAddress(
-      accounts[0].address,
-    )
-  })
-}
-
-const parentCannotBurnParentControlledFuses = ({
-  fixture,
-}: BaseTestParameters) => {
-  it('Parent cannot burn parent-controlled fuses', async () => {
-    const { nameWrapper } = await loadFixture(fixture)
-
-    await expect(nameWrapper)
-      .write('setChildFuses', [parentNode, childLabelHash, 1 << 18, 0n])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-  })
-}
-
-const ownerIsOwnerWhenExpired = ({ fixture }: BaseTestParameters) => {
-  it('Owner is still owner when expired', async () => {
-    const { nameWrapper, accounts, publicClient } = await loadFixture(fixture)
-
-    const timestamp = await publicClient.getBlock().then((b) => b.timestamp)
-    const [, , expiry] = await nameWrapper.read.getData([childNodeId])
-
-    expect(expiry).toBeLessThan(timestamp)
-
-    await expect(
-      nameWrapper.read.ownerOf([childNodeId]),
-    ).resolves.toEqualAddress(accounts[1].address)
-  })
-}
-
-const ownerCannotBurnFuses = ({ fixture }: BaseTestParameters) => {
-  it('Owner cannot burn CU because PCC is not burned', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
-
-    await expect(nameWrapper)
-      .write('setFuses', [childNode, CANNOT_UNWRAP], { account: accounts[1] })
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-  })
-
-  it('Owner cannot burn other fuses because CU and PCC are not burned', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
-
-    await expect(nameWrapper)
-      .write('setFuses', [childNode, CANNOT_SET_RESOLVER], {
-        account: accounts[1],
-      })
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-  })
-}
-
-const ownerCanUnwrap = ({ fixture }: BaseTestParameters) => {
-  it('Owner can unwrap', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
-
-    await nameWrapper.write.unwrap(
-      [parentNode, childLabelHash, accounts[1].address],
-      { account: accounts[1] },
-    )
-
-    await expect(
-      nameWrapper.read.ownerOf([childNodeId]),
-    ).resolves.toEqualAddress(zeroAddress)
-  })
-}
-
-const parentCanBurnParentControlledFusesWithExpiry = ({
-  fixture,
-}: BaseTestParameters) => {
-  it('Parent cannot burn parent-controlled fuses as they reset to 0', async () => {
-    const { nameWrapper } = await loadFixture(fixture)
-
-    await nameWrapper.write.setChildFuses([
-      parentNode,
-      childLabelHash,
-      1 << 18,
-      0n,
-    ])
-
-    // expired names get normalised to 0
-    const [, fuses] = await nameWrapper.read.getData([childNodeId])
-    expect(fuses).toEqual(0)
-  })
-
-  it('Parent can burn parent-controlled fuses, if expiry is extended', async () => {
-    const { nameWrapper } = await loadFixture(fixture)
-
-    await nameWrapper.write.setChildFuses([
-      parentNode,
-      childLabelHash,
-      1 << 18,
-      MAX_EXPIRY,
-    ])
-
-    const [, fuses] = await nameWrapper.read.getData([childNodeId])
-    expect(fuses).toEqual(1 << 18)
-  })
-}
-
-const parentCanBurnFusesOrPCC = ({ fixture }: BaseTestParameters) => {
-  it('Parent can burn fuses with setChildFuses()', async () => {
-    const { nameWrapper } = await loadFixture(fixture)
-
-    await nameWrapper.write.setChildFuses([
-      parentNode,
-      childLabelHash,
-      CANNOT_UNWRAP | PARENT_CANNOT_CONTROL | CANNOT_SET_RESOLVER,
-      0n,
-    ])
-
-    const [, fuses] = await nameWrapper.read.getData([childNodeId])
-    expect(fuses).toEqual(
-      CANNOT_UNWRAP | PARENT_CANNOT_CONTROL | CANNOT_SET_RESOLVER,
-    )
-  })
-
-  it('Parent can burn fuses with setSubnodeOwner()', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
-
-    await nameWrapper.write.setSubnodeOwner([
-      parentNode,
-      childLabel,
-      accounts[1].address,
-      CANNOT_UNWRAP | PARENT_CANNOT_CONTROL | CANNOT_SET_RESOLVER,
-      0n,
-    ])
-
-    const [, fuses] = await nameWrapper.read.getData([childNodeId])
-    expect(fuses).toEqual(
-      CANNOT_UNWRAP | PARENT_CANNOT_CONTROL | CANNOT_SET_RESOLVER,
-    )
-  })
-
-  it('Parent can burn fuses with setSubnodeRecord()', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
-
-    await nameWrapper.write.setSubnodeRecord([
-      parentNode,
-      childLabel,
-      accounts[1].address,
-      zeroAddress,
-      0n,
-      CANNOT_UNWRAP | PARENT_CANNOT_CONTROL | CANNOT_SET_RESOLVER,
-      0n,
-    ])
-
-    const [, fuses] = await nameWrapper.read.getData([childNodeId])
-    expect(fuses).toEqual(
-      CANNOT_UNWRAP | PARENT_CANNOT_CONTROL | CANNOT_SET_RESOLVER,
-    )
-  })
-
-  it('Parent cannot burn fuses if PCC is not burnt too', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
-
-    await expect(nameWrapper)
-      .write('setSubnodeRecord', [
-        parentNode,
-        childLabel,
-        accounts[1].address,
-        zeroAddress,
-        0n,
-        CANNOT_UNWRAP | CANNOT_SET_RESOLVER,
-        0n,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-
-    await expect(nameWrapper)
-      .write('setChildFuses', [
-        parentNode,
-        childLabelHash,
-        CANNOT_UNWRAP | CANNOT_SET_RESOLVER,
-        0n,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-
-    await expect(nameWrapper)
-      .write('setSubnodeOwner', [
-        parentNode,
-        childLabel,
-        accounts[1].address,
-        CANNOT_UNWRAP | CANNOT_SET_RESOLVER,
-        0n,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-  })
-}
-
-const parentCanBurnParentControlledFuses = ({
-  fixture,
-}: BaseTestParameters) => {
-  it('Parent can burn parent-controlled fuses', async () => {
-    const { nameWrapper } = await loadFixture(fixture)
-
-    await nameWrapper.write.setChildFuses([
-      parentNode,
-      childLabelHash,
-      1 << 18,
-      0n,
-    ])
-
-    const [, fuses] = await nameWrapper.read.getData([childNodeId])
-    expect(fuses).toEqual(1 << 18)
-  })
-}
-
-const testStateTransition1000to1010 = ({ fixture }: BaseTestParameters) => {
-  it('1000 => 1010 - Parent cannot burn PCC with setChildFuses()', async () => {
-    const { nameWrapper } = await loadFixture(fixture)
-
-    await expect(nameWrapper)
-      .write('setChildFuses', [
+  const parentCannotBurnFusesWhenPCCisBurned = ({
+    fixture,
+  }: BaseTestParameters) => {
+    it('Parent cannot burn fuses with setChildFuses()', async () => {
+      const { nameWrapper } = await loadFixture(fixture)
+
+      const [, fusesBefore] = await nameWrapper.read.getData([childNodeId])
+      expect(fusesBefore).toEqual(PARENT_CANNOT_CONTROL)
+
+      await expect(
+        nameWrapper.write.setChildFuses([
+          parentNode,
+          childLabelHash,
+          CANNOT_UNWRAP | PARENT_CANNOT_CONTROL | CANNOT_SET_RESOLVER,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
+
+      await expect(
+        nameWrapper.write.setChildFuses([
+          parentNode,
+          childLabelHash,
+          CANNOT_UNWRAP | PARENT_CANNOT_CONTROL,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
+
+      // parent can burn PCC again, but has no effect since it's already burnt
+      await nameWrapper.write.setChildFuses([
         parentNode,
         childLabelHash,
         PARENT_CANNOT_CONTROL,
         0n,
       ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-  })
 
-  it('1000 => 1010 - Parent cannot burn PCC with setSubnodeOwner()', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
+      const [, fuses] = await nameWrapper.read.getData([childNodeId])
+      expect(fuses).toEqual(PARENT_CANNOT_CONTROL)
+    })
 
-    await expect(nameWrapper)
-      .write('setSubnodeOwner', [
-        parentNode,
-        childLabel,
-        accounts[1].address,
-        PARENT_CANNOT_CONTROL,
-        0n,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-  })
+    it('Parent cannot burn fuses with setSubnodeOwner()', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
 
-  it('1000 => 1010 - Parent cannot burn PCC with setSubnodeRecord()', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
+      await expect(
+        nameWrapper.write.setSubnodeOwner([
+          parentNode,
+          childLabel,
+          accounts[1].address,
+          CANNOT_SET_RESOLVER,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
 
-    await expect(nameWrapper)
-      .write('setSubnodeRecord', [
-        parentNode,
-        childLabel,
-        accounts[1].address,
-        zeroAddress,
-        0n,
-        PARENT_CANNOT_CONTROL,
-        0n,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-  })
-}
+      await expect(
+        nameWrapper.write.setSubnodeOwner([
+          parentNode,
+          childLabel,
+          accounts[1].address,
+          CANNOT_UNWRAP | CANNOT_SET_RESOLVER,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
 
-const parentCanExtendWithSetChildFuses = ({ fixture }: BaseTestParameters) => {
-  it('Child should have a <parent expiry before extending', async () => {
-    const { nameWrapper, baseRegistrar, publicClient } = await loadFixture(
-      fixture,
-    )
+      await expect(
+        nameWrapper.write.setSubnodeOwner([
+          parentNode,
+          childLabel,
+          accounts[1].address,
+          PARENT_CANNOT_CONTROL | CANNOT_UNWRAP | CANNOT_SET_RESOLVER,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
+    })
 
-    const [, , childExpiry] = await nameWrapper.read.getData([childNodeId])
-    const parentExpiry = await baseRegistrar.read.nameExpires([parentLabelId])
-    expect(childExpiry).toBeLessThan(parentExpiry)
+    it('Parent cannot burn fuses with setSubnodeRecord()', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
 
-    const timestamp = await publicClient.getBlock().then((b) => b.timestamp)
-    expect(childExpiry).toBeGreaterThan(timestamp)
-  })
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
+          parentNode,
+          childLabel,
+          accounts[1].address,
+          zeroAddress,
+          0n,
+          CANNOT_UNWRAP | PARENT_CANNOT_CONTROL | CANNOT_SET_RESOLVER,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
 
-  it('Parent can extend expiry with setChildFuses()', async () => {
-    const { nameWrapper, baseRegistrar } = await loadFixture(fixture)
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
+          parentNode,
+          childLabel,
+          accounts[1].address,
+          zeroAddress,
+          0n,
+          CANNOT_UNWRAP | PARENT_CANNOT_CONTROL,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
 
-    const parentExpiry = await baseRegistrar.read.nameExpires([parentLabelId])
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
+          parentNode,
+          childLabel,
+          accounts[1].address,
+          zeroAddress,
+          0n,
+          PARENT_CANNOT_CONTROL,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
+    })
+  }
 
-    await nameWrapper.write.setChildFuses([
-      parentNode,
-      childLabelHash,
-      CAN_DO_EVERYTHING,
-      MAX_EXPIRY,
-    ])
+  const parentCannotReplaceOwner = ({ fixture }: BaseTestParameters) => {
+    it('Parent cannot replace owner with setSubnodeOwner()', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
 
-    const [, , expiry] = await nameWrapper.read.getData([childNodeId])
+      await expect(
+        nameWrapper.read.ownerOf([childNodeId]),
+      ).resolves.toEqualAddress(accounts[1].address)
 
-    expect(expiry).toEqual(parentExpiry + GRACE_PERIOD)
-  })
+      await expect(
+        nameWrapper.write.setSubnodeOwner([
+          parentNode,
+          childLabel,
+          accounts[0].address,
+          CAN_DO_EVERYTHING,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
 
-  it('Parent cannot extend expiry with setSubnodeOwner()', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
+      await expect(
+        nameWrapper.read.ownerOf([childNodeId]),
+      ).resolves.toEqualAddress(accounts[1].address)
+    })
 
-    await expect(nameWrapper)
-      .write('setSubnodeOwner', [
-        parentNode,
-        childLabel,
-        accounts[1].address,
-        CAN_DO_EVERYTHING,
-        MAX_EXPIRY,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-  })
+    it('Parent cannot replace owner with setSubnodeRecord()', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
 
-  it('Parent cannot extend expiry with setSubnodeRecord()', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
+      await expect(
+        nameWrapper.read.ownerOf([childNodeId]),
+      ).resolves.toEqualAddress(accounts[1].address)
 
-    await expect(nameWrapper)
-      .write('setSubnodeRecord', [
-        parentNode,
-        childLabel,
-        accounts[1].address,
-        zeroAddress,
-        0n,
-        CAN_DO_EVERYTHING,
-        MAX_EXPIRY,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-  })
-}
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
+          parentNode,
+          childLabel,
+          accounts[0].address,
+          zeroAddress,
+          0n,
+          CAN_DO_EVERYTHING,
+          0n,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([childNode])
 
-const parentCannotBurnFusesWhenPCCisBurned = ({
-  fixture,
-}: BaseTestParameters) => {
-  it('Parent cannot burn fuses with setChildFuses()', async () => {
-    const { nameWrapper } = await loadFixture(fixture)
+      await expect(
+        nameWrapper.read.ownerOf([childNodeId]),
+      ).resolves.toEqualAddress(accounts[1].address)
+    })
+  }
 
-    const [, fusesBefore] = await nameWrapper.read.getData([childNodeId])
-    expect(fusesBefore).toEqual(PARENT_CANNOT_CONTROL)
+  const parentCannotUnwrapChild = ({ fixture }: BaseTestParameters) => {
+    it('Parent cannot unwrap itself', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
 
-    await expect(nameWrapper)
-      .write('setChildFuses', [
-        parentNode,
-        childLabelHash,
-        CANNOT_UNWRAP | PARENT_CANNOT_CONTROL | CANNOT_SET_RESOLVER,
-        0n,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
+      await expect(
+        nameWrapper.write.unwrapETH2LD([
+          parentLabelHash,
+          accounts[0].address,
+          accounts[0].address,
+        ]),
+      )
+        .toBeRevertedWithCustomError('OperationProhibited')
+        .withArgs([parentNode])
+    })
 
-    await expect(nameWrapper)
-      .write('setChildFuses', [
-        parentNode,
-        childLabelHash,
-        CANNOT_UNWRAP | PARENT_CANNOT_CONTROL,
-        0n,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
+    it('Parent cannot unwrap child', async () => {
+      const { nameWrapper, accounts } = await loadFixture(fixture)
 
-    // parent can burn PCC again, but has no effect since it's already burnt
-    await nameWrapper.write.setChildFuses([
-      parentNode,
-      childLabelHash,
-      PARENT_CANNOT_CONTROL,
-      0n,
-    ])
+      await expect(
+        nameWrapper.write.unwrap([
+          parentNode,
+          childLabelHash,
+          accounts[0].address,
+        ]),
+      )
+        .toBeRevertedWithCustomError('Unauthorised')
+        .withArgs([childNode, getAddress(accounts[0].address)])
+    })
 
-    const [, fuses] = await nameWrapper.read.getData([childNodeId])
-    expect(fuses).toEqual(PARENT_CANNOT_CONTROL)
-  })
+    it('Parent cannot call ens.setSubnodeOwner() to forcefully unwrap', async () => {
+      const { ensRegistry, accounts } = await loadFixture(fixture)
 
-  it('Parent cannot burn fuses with setSubnodeOwner()', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
+      await expect(
+        ensRegistry.write.setSubnodeOwner([
+          parentNode,
+          childNode,
+          accounts[0].address,
+        ]),
+      ).toBeRevertedWithoutReason()
+    })
+  }
 
-    await expect(nameWrapper)
-      .write('setSubnodeOwner', [
-        parentNode,
-        childLabel,
-        accounts[1].address,
-        CANNOT_SET_RESOLVER,
-        0n,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
+  const ownerResetsToZeroWhenExpired = ({
+    fixture,
+    expectedFuses,
+  }: BaseTestParameters & { expectedFuses: number }) => {
+    it('Owner resets to 0 after expiry', async () => {
+      const { nameWrapper, accounts, publicClient, testClient } =
+        await loadFixture(fixture)
 
-    await expect(nameWrapper)
-      .write('setSubnodeOwner', [
-        parentNode,
-        childLabel,
-        accounts[1].address,
-        CANNOT_UNWRAP | CANNOT_SET_RESOLVER,
-        0n,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
+      const [ownerBefore, fusesBefore, expiryBefore] =
+        await nameWrapper.read.getData([childNodeId])
+      const timestampBefore = await publicClient
+        .getBlock()
+        .then((b) => b.timestamp)
+      // not expired
+      expect(ownerBefore).toEqualAddress(accounts[1].address)
+      expect(fusesBefore).toEqual(expectedFuses)
+      expect(expiryBefore).toBeGreaterThan(timestampBefore)
 
-    await expect(nameWrapper)
-      .write('setSubnodeOwner', [
-        parentNode,
-        childLabel,
-        accounts[1].address,
-        PARENT_CANNOT_CONTROL | CANNOT_UNWRAP | CANNOT_SET_RESOLVER,
-        0n,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-  })
+      // force expiry
+      await testClient.increaseTime({ seconds: Number(2n * DAY) })
+      await testClient.mine({ blocks: 1 })
 
-  it('Parent cannot burn fuses with setSubnodeRecord()', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
+      const [ownerAfter, fusesAfter, expiryAfter] =
+        await nameWrapper.read.getData([childNodeId])
+      const timestampAfter = await publicClient
+        .getBlock()
+        .then((b) => b.timestamp)
+      // owner and fuses are reset when expired
+      expect(ownerAfter).toEqualAddress(zeroAddress)
+      expect(fusesAfter).toEqual(0)
+      expect(expiryAfter).toBeLessThan(timestampAfter)
+    })
+  }
 
-    await expect(nameWrapper)
-      .write('setSubnodeRecord', [
-        parentNode,
-        childLabel,
-        accounts[1].address,
-        zeroAddress,
-        0n,
-        CANNOT_UNWRAP | PARENT_CANNOT_CONTROL | CANNOT_SET_RESOLVER,
-        0n,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-
-    await expect(nameWrapper)
-      .write('setSubnodeRecord', [
-        parentNode,
-        childLabel,
-        accounts[1].address,
-        zeroAddress,
-        0n,
-        CANNOT_UNWRAP | PARENT_CANNOT_CONTROL,
-        0n,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-
-    await expect(nameWrapper)
-      .write('setSubnodeRecord', [
-        parentNode,
-        childLabel,
-        accounts[1].address,
-        zeroAddress,
-        0n,
-        PARENT_CANNOT_CONTROL,
-        0n,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-  })
-}
-
-const parentCannotReplaceOwner = ({ fixture }: BaseTestParameters) => {
-  it('Parent cannot replace owner with setSubnodeOwner()', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
-
-    await expect(
-      nameWrapper.read.ownerOf([childNodeId]),
-    ).resolves.toEqualAddress(accounts[1].address)
-
-    await expect(nameWrapper)
-      .write('setSubnodeOwner', [
-        parentNode,
-        childLabel,
-        accounts[0].address,
-        CAN_DO_EVERYTHING,
-        0n,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-
-    await expect(
-      nameWrapper.read.ownerOf([childNodeId]),
-    ).resolves.toEqualAddress(accounts[1].address)
-  })
-
-  it('Parent cannot replace owner with setSubnodeRecord()', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
-
-    await expect(
-      nameWrapper.read.ownerOf([childNodeId]),
-    ).resolves.toEqualAddress(accounts[1].address)
-
-    await expect(nameWrapper)
-      .write('setSubnodeRecord', [
-        parentNode,
-        childLabel,
-        accounts[0].address,
-        zeroAddress,
-        0n,
-        CAN_DO_EVERYTHING,
-        0n,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(childNode)
-
-    await expect(
-      nameWrapper.read.ownerOf([childNodeId]),
-    ).resolves.toEqualAddress(accounts[1].address)
-  })
-}
-
-const parentCannotUnwrapChild = ({ fixture }: BaseTestParameters) => {
-  it('Parent cannot unwrap itself', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
-
-    await expect(nameWrapper)
-      .write('unwrapETH2LD', [
-        parentLabelHash,
-        accounts[0].address,
-        accounts[0].address,
-      ])
-      .toBeRevertedWithCustomError('OperationProhibited')
-      .withArgs(parentNode)
-  })
-
-  it('Parent cannot unwrap child', async () => {
-    const { nameWrapper, accounts } = await loadFixture(fixture)
-
-    await expect(nameWrapper)
-      .write('unwrap', [parentNode, childLabelHash, accounts[0].address])
-      .toBeRevertedWithCustomError('Unauthorised')
-      .withArgs(childNode, getAddress(accounts[0].address))
-  })
-
-  it('Parent cannot call ens.setSubnodeOwner() to forcefully unwrap', async () => {
-    const { ensRegistry, accounts } = await loadFixture(fixture)
-
-    await expect(ensRegistry)
-      .write('setSubnodeOwner', [parentNode, childNode, accounts[0].address])
-      .toBeRevertedWithoutReason()
-  })
-}
-
-const ownerResetsToZeroWhenExpired = ({
-  fixture,
-  expectedFuses,
-}: BaseTestParameters & { expectedFuses: number }) => {
-  it('Owner resets to 0 after expiry', async () => {
-    const { nameWrapper, accounts, publicClient, testClient } =
-      await loadFixture(fixture)
-
-    const [ownerBefore, fusesBefore, expiryBefore] =
-      await nameWrapper.read.getData([childNodeId])
-    const timestampBefore = await publicClient
-      .getBlock()
-      .then((b) => b.timestamp)
-    // not expired
-    expect(ownerBefore).toEqualAddress(accounts[1].address)
-    expect(fusesBefore).toEqual(expectedFuses)
-    expect(expiryBefore).toBeGreaterThan(timestampBefore)
-
-    // force expiry
-    await testClient.increaseTime({ seconds: Number(2n * DAY) })
-    await testClient.mine({ blocks: 1 })
-
-    const [ownerAfter, fusesAfter, expiryAfter] =
-      await nameWrapper.read.getData([childNodeId])
-    const timestampAfter = await publicClient
-      .getBlock()
-      .then((b) => b.timestamp)
-    // owner and fuses are reset when expired
-    expect(ownerAfter).toEqualAddress(zeroAddress)
-    expect(fusesAfter).toEqual(0)
-    expect(expiryAfter).toBeLessThan(timestampAfter)
-  })
-}
-
-export const shouldRespectConstraints = () => {
   describe("0000 - Wrapped expired without CU/PCC burned, Parent's CU not burned", () => {
     const fixture = setupState0000DW
 
@@ -955,15 +1011,16 @@ export const shouldRespectConstraints = () => {
       const [, fusesBefore] = await nameWrapper.read.getData([childNodeId])
       expect(fusesBefore).toEqual(0)
 
-      await expect(nameWrapper)
-        .write('setChildFuses', [
+      await expect(
+        nameWrapper.write.setChildFuses([
           parentNode,
           childLabelHash,
           CANNOT_UNWRAP | PARENT_CANNOT_CONTROL | CANNOT_SET_RESOLVER,
           0n,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
     })
 
     parentCanReplaceOwner({ fixture })
@@ -1028,14 +1085,15 @@ export const shouldRespectConstraints = () => {
     it('Parent cannot unwrap itself', async () => {
       const { nameWrapper, accounts } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('unwrapETH2LD', [
+      await expect(
+        nameWrapper.write.unwrapETH2LD([
           parentLabelHash,
           accounts[0].address,
           accounts[0].address,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(parentNode)
+        .withArgs([parentNode])
     })
 
     ownerCannotBurnFuses({ fixture })
@@ -1052,15 +1110,16 @@ export const shouldRespectConstraints = () => {
     it('0000 => 0010 - Parent cannot burn PCC with setChildFuses()', async () => {
       const { nameWrapper } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setChildFuses', [
+      await expect(
+        nameWrapper.write.setChildFuses([
           parentNode,
           childLabelHash,
           PARENT_CANNOT_CONTROL,
           0n,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
     })
   })
 
@@ -1090,19 +1149,28 @@ export const shouldRespectConstraints = () => {
     it('0000 => 0100 - DW => CU Parent - cannot burn CANNOT_UNWRAP with setChildFuses()', async () => {
       const { nameWrapper } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setChildFuses', [parentNode, childLabelHash, CANNOT_UNWRAP, 0n])
+      await expect(
+        nameWrapper.write.setChildFuses([
+          parentNode,
+          childLabelHash,
+          CANNOT_UNWRAP,
+          0n,
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
     })
 
     it('0000 => 0100 - DW => CU - Owner cannot burn CANNOT_UNWRAP with setFuses()', async () => {
       const { nameWrapper, accounts } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setFuses', [childNode, CANNOT_UNWRAP], { account: accounts[1] })
+      await expect(
+        nameWrapper.write.setFuses([childNode, CANNOT_UNWRAP], {
+          account: accounts[1],
+        }),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
     })
   })
 
@@ -1112,19 +1180,28 @@ export const shouldRespectConstraints = () => {
     it('0001 => 0101 - PCU => CU_PCU - Parent cannot burn CANNOT_UNWRAP with setChildFuses()', async () => {
       const { nameWrapper } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setChildFuses', [parentNode, childLabelHash, CANNOT_UNWRAP, 0n])
+      await expect(
+        nameWrapper.write.setChildFuses([
+          parentNode,
+          childLabelHash,
+          CANNOT_UNWRAP,
+          0n,
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
     })
 
     it('0001 => 0101 - PCU => CU_PCU -  Owner cannot burn CANNOT_UNWRAP with setFuses()', async () => {
       const { nameWrapper, accounts } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setFuses', [childNode, CANNOT_UNWRAP], { account: accounts[1] })
+      await expect(
+        nameWrapper.write.setFuses([childNode, CANNOT_UNWRAP], {
+          account: accounts[1],
+        }),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
     })
   })
 
@@ -1134,15 +1211,16 @@ export const shouldRespectConstraints = () => {
     it('0000 => 0010 - DW => PCC - Parent cannot burn PARENT_CANNOT_CONTROL with setChildFuses()', async () => {
       const { nameWrapper } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setChildFuses', [
+      await expect(
+        nameWrapper.write.setChildFuses([
           parentNode,
           childLabelHash,
           PARENT_CANNOT_CONTROL,
           0n,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
     })
   })
 
@@ -1216,16 +1294,17 @@ export const shouldRespectConstraints = () => {
     it('Parent cannot unburn fuses with setSubnodeOwner()', async () => {
       const { nameWrapper, accounts } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setSubnodeOwner', [
+      await expect(
+        nameWrapper.write.setSubnodeOwner([
           parentNode,
           childLabel,
           accounts[1].address,
           0,
           0n,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
 
       const [, fuses] = await nameWrapper.read.getData([childNodeId])
       expect(fuses).toEqual(PARENT_CANNOT_CONTROL)
@@ -1234,8 +1313,8 @@ export const shouldRespectConstraints = () => {
     it('Parent cannot unburn fuses with setSubnodeRecord()', async () => {
       const { nameWrapper, accounts } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setSubnodeRecord', [
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
           parentNode,
           childLabel,
           accounts[1].address,
@@ -1243,9 +1322,10 @@ export const shouldRespectConstraints = () => {
           0n,
           0,
           0n,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
 
       const [, fuses] = await nameWrapper.read.getData([childNodeId])
       expect(fuses).toEqual(PARENT_CANNOT_CONTROL)
@@ -1269,12 +1349,13 @@ export const shouldRespectConstraints = () => {
     it('Owner cannot burn fuses because CU is unburned', async () => {
       const { nameWrapper, accounts } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setFuses', [childNode, CANNOT_SET_RESOLVER], {
+      await expect(
+        nameWrapper.write.setFuses([childNode, CANNOT_SET_RESOLVER], {
           account: accounts[1],
-        })
+        }),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
     })
 
     it('Owner cannot unwrap and wrap to unburn PCC', async () => {
@@ -1316,32 +1397,39 @@ export const shouldRespectConstraints = () => {
     it('1000 => 1100 - NE => NE_CU - Parent cannot burn CU with setChildFuses()', async () => {
       const { nameWrapper } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setChildFuses', [parentNode, childLabelHash, CANNOT_UNWRAP, 0n])
+      await expect(
+        nameWrapper.write.setChildFuses([
+          parentNode,
+          childLabelHash,
+          CANNOT_UNWRAP,
+          0n,
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
     })
 
     it('1000 => 1100 - NE => NE_CU - Parent cannot burn CU with setSubnodeOwner()', async () => {
       const { nameWrapper, accounts } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setSubnodeOwner', [
+      await expect(
+        nameWrapper.write.setSubnodeOwner([
           parentNode,
           childLabel,
           accounts[1].address,
           CANNOT_UNWRAP,
           0n,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
     })
 
     it('1000 => 1100 - NE => NE_CU - Parent cannot burn CU with setSubnodeRecord()', async () => {
       const { nameWrapper, accounts } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setSubnodeRecord', [
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
           parentNode,
           childLabel,
           accounts[1].address,
@@ -1349,18 +1437,22 @@ export const shouldRespectConstraints = () => {
           0n,
           CANNOT_UNWRAP,
           0n,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
     })
 
     it('1000 => 1100 - NE => NE_CU - Owner cannot burn CU with setFuses()', async () => {
       const { nameWrapper, accounts } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setFuses', [childNode, CANNOT_UNWRAP], { account: accounts[1] })
+      await expect(
+        nameWrapper.write.setFuses([childNode, CANNOT_UNWRAP], {
+          account: accounts[1],
+        }),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
     })
   })
 
@@ -1370,32 +1462,39 @@ export const shouldRespectConstraints = () => {
     it('1001 => 1101 - NE_PCU => NE_CU_PCU -  Parent cannot burn CU with setChildFuses()', async () => {
       const { nameWrapper } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setChildFuses', [parentNode, childLabelHash, CANNOT_UNWRAP, 0n])
+      await expect(
+        nameWrapper.write.setChildFuses([
+          parentNode,
+          childLabelHash,
+          CANNOT_UNWRAP,
+          0n,
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
     })
 
     it('1001 => 1101 - NE_PCU => NE_CU_PCU - Parent cannot burn CU with setSubnodeOwner()', async () => {
       const { nameWrapper, accounts } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setSubnodeOwner', [
+      await expect(
+        nameWrapper.write.setSubnodeOwner([
           parentNode,
           childLabel,
           accounts[1].address,
           CANNOT_UNWRAP,
           0n,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
     })
 
     it('1001 => 1101 - NE_PCU => NE_CU_PCU - Parent cannot burn CU with setSubnodeRecord()', async () => {
       const { nameWrapper, accounts } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setSubnodeRecord', [
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
           parentNode,
           childLabel,
           accounts[1].address,
@@ -1403,18 +1502,22 @@ export const shouldRespectConstraints = () => {
           0n,
           CANNOT_UNWRAP,
           0n,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
     })
 
     it('1001 => 1101 - NE_PCU => NE_CU_PCU - Owner cannot burn CU with setFuses()', async () => {
       const { nameWrapper, accounts } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setFuses', [childNode, CANNOT_UNWRAP], { account: accounts[1] })
+      await expect(
+        nameWrapper.write.setFuses([childNode, CANNOT_UNWRAP], {
+          account: accounts[1],
+        }),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
     })
   })
 
@@ -1440,16 +1543,17 @@ export const shouldRespectConstraints = () => {
     it('Parent cannot unburn fuses with setSubnodeOwner()', async () => {
       const { nameWrapper, accounts } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setSubnodeOwner', [
+      await expect(
+        nameWrapper.write.setSubnodeOwner([
           parentNode,
           childLabel,
           accounts[1].address,
           0,
           0n,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
 
       const [, fuses] = await nameWrapper.read.getData([childNodeId])
       expect(fuses).toEqual(PARENT_CANNOT_CONTROL | CANNOT_UNWRAP)
@@ -1458,8 +1562,8 @@ export const shouldRespectConstraints = () => {
     it('Parent cannot unburn fuses with setSubnodeRecord()', async () => {
       const { nameWrapper, accounts } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('setSubnodeRecord', [
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
           parentNode,
           childLabel,
           accounts[1].address,
@@ -1467,9 +1571,10 @@ export const shouldRespectConstraints = () => {
           0n,
           0,
           0n,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
 
       const [, fuses] = await nameWrapper.read.getData([childNodeId])
       expect(fuses).toEqual(PARENT_CANNOT_CONTROL | CANNOT_UNWRAP)
@@ -1510,12 +1615,16 @@ export const shouldRespectConstraints = () => {
     it('Owner cannot unwrap', async () => {
       const { nameWrapper, accounts } = await loadFixture(fixture)
 
-      await expect(nameWrapper)
-        .write('unwrap', [parentNode, childLabelHash, accounts[1].address], {
-          account: accounts[1],
-        })
+      await expect(
+        nameWrapper.write.unwrap(
+          [parentNode, childLabelHash, accounts[1].address],
+          {
+            account: accounts[1],
+          },
+        ),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(childNode)
+        .withArgs([childNode])
     })
 
     ownerResetsToZeroWhenExpired({
