@@ -1,20 +1,23 @@
-import { loadFixture } from '@nomicfoundation/hardhat-toolbox-viem/network-helpers.js'
-import { expect } from 'chai'
+import type { NetworkConnection } from 'hardhat/types/network'
 import { getAddress, namehash, zeroAddress } from 'viem'
+
 import {
   CANNOT_SET_RESOLVER,
   CANNOT_UNWRAP,
   expectOwnerOf,
-  deployNameWrapperWithUtils as fixture,
+  type LoadNameWrapperFixture,
 } from '../fixtures/utils.js'
 
-export const setResolverTests = () => {
+export const setResolverTests = (
+  connection: NetworkConnection,
+  loadNameWrapperFixture: LoadNameWrapperFixture,
+) => {
   describe('setResolver', () => {
     const label = 'setresolver'
     const name = `${label}.eth`
 
-    async function setResolverFixture() {
-      const initial = await loadFixture(fixture)
+    async function fixture() {
+      const initial = await loadNameWrapperFixture()
       const { actions } = initial
 
       await actions.registerSetupAndWrapName({
@@ -25,8 +28,11 @@ export const setResolverTests = () => {
       return initial
     }
 
+    const loadFixture = async () =>
+      connection.networkHelpers.loadFixture(fixture)
+
     it('Can be called by the owner', async () => {
-      const { nameWrapper, accounts } = await loadFixture(setResolverFixture)
+      const { nameWrapper, accounts } = await loadFixture()
 
       await expectOwnerOf(name).on(nameWrapper).toBe(accounts[0])
 
@@ -34,9 +40,7 @@ export const setResolverTests = () => {
     })
 
     it('Performs the appropriate function on the ENS registry.', async () => {
-      const { ensRegistry, nameWrapper, accounts } = await loadFixture(
-        setResolverFixture,
-      )
+      const { ensRegistry, nameWrapper, accounts } = await loadFixture()
 
       await expect(
         ensRegistry.read.resolver([namehash(name)]),
@@ -50,7 +54,7 @@ export const setResolverTests = () => {
     })
 
     it('Can be called by an account authorised by the owner.', async () => {
-      const { nameWrapper, accounts } = await loadFixture(setResolverFixture)
+      const { nameWrapper, accounts } = await loadFixture()
 
       await expectOwnerOf(name).on(nameWrapper).toBe(accounts[0])
 
@@ -65,25 +69,27 @@ export const setResolverTests = () => {
     })
 
     it('Cannot be called by anyone else.', async () => {
-      const { nameWrapper, accounts } = await loadFixture(setResolverFixture)
+      const { nameWrapper, accounts } = await loadFixture()
 
-      await expect(nameWrapper)
-        .write('setResolver', [namehash(name), accounts[1].address], {
+      await expect(
+        nameWrapper.write.setResolver([namehash(name), accounts[1].address], {
           account: accounts[1],
-        })
+        }),
+      )
         .toBeRevertedWithCustomError('Unauthorised')
-        .withArgs(namehash(name), getAddress(accounts[1].address))
+        .withArgs([namehash(name), getAddress(accounts[1].address)])
     })
 
     it('Cannot be called if CANNOT_SET_RESOLVER is burned', async () => {
-      const { nameWrapper, accounts } = await loadFixture(setResolverFixture)
+      const { nameWrapper, accounts } = await loadFixture()
 
       await nameWrapper.write.setFuses([namehash(name), CANNOT_SET_RESOLVER])
 
-      await expect(nameWrapper)
-        .write('setResolver', [namehash(name), accounts[1].address])
+      await expect(
+        nameWrapper.write.setResolver([namehash(name), accounts[1].address]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(namehash(name))
+        .withArgs([namehash(name)])
     })
   })
 }
