@@ -1,7 +1,6 @@
-import { loadFixture } from '@nomicfoundation/hardhat-toolbox-viem/network-helpers.js'
-import { expect } from 'chai'
-import hre from 'hardhat'
+import type { NetworkConnection } from 'hardhat/types/network'
 import { getAddress, labelhash, namehash, zeroAddress } from 'viem'
+
 import { DAY } from '../../fixtures/constants.js'
 import { dnsEncodeName } from '../../fixtures/dnsEncodeName.js'
 import { toLabelId, toNameId } from '../../fixtures/utils.js'
@@ -15,19 +14,22 @@ import {
   MAX_EXPIRY,
   PARENT_CANNOT_CONTROL,
   expectOwnerOf,
-  deployNameWrapperWithUtils as fixture,
   zeroAccount,
+  type LoadNameWrapperFixture,
 } from '../fixtures/utils.js'
 
-export const setSubnodeRecordTests = () =>
+export const setSubnodeRecordTests = (
+  connection: NetworkConnection,
+  loadNameWrapperFixture: LoadNameWrapperFixture,
+) =>
   describe('setSubnodeRecord()', () => {
     const label = 'subdomain2'
     const sublabel = 'sub'
     const name = `${label}.eth`
     const subname = `${sublabel}.${name}`
 
-    async function setSubnodeRecordFixture() {
-      const initial = await loadFixture(fixture)
+    async function fixture() {
+      const initial = await loadNameWrapperFixture()
       const { actions, accounts } = initial
 
       await actions.registerSetupAndWrapName({
@@ -37,10 +39,12 @@ export const setSubnodeRecordTests = () =>
 
       return { ...initial, resolverAddress: accounts[0].address }
     }
+    const loadFixture = async () =>
+      connection.networkHelpers.loadFixture(fixture)
 
     it('Can be called by the owner of a name', async () => {
       const { ensRegistry, nameWrapper, actions, accounts, resolverAddress } =
-        await loadFixture(setSubnodeRecordFixture)
+        await loadFixture()
 
       await expectOwnerOf(name).on(nameWrapper).toBe(accounts[0])
 
@@ -60,7 +64,7 @@ export const setSubnodeRecordTests = () =>
 
     it('Can be called by an account authorised by the owner.', async () => {
       const { ensRegistry, nameWrapper, actions, accounts, resolverAddress } =
-        await loadFixture(setSubnodeRecordFixture)
+        await loadFixture()
 
       await expectOwnerOf(name).on(nameWrapper).toBe(accounts[0])
 
@@ -83,7 +87,7 @@ export const setSubnodeRecordTests = () =>
 
     it('Transfers the wrapped token to the target address.', async () => {
       const { nameWrapper, actions, accounts, resolverAddress } =
-        await loadFixture(setSubnodeRecordFixture)
+        await loadFixture()
 
       await actions.setSubnodeRecord.onNameWrapper({
         parentName: name,
@@ -99,12 +103,10 @@ export const setSubnodeRecordTests = () =>
     })
 
     it('Will not allow wrapping with a target address of 0x0', async () => {
-      const { nameWrapper, resolverAddress } = await loadFixture(
-        setSubnodeRecordFixture,
-      )
+      const { nameWrapper, resolverAddress } = await loadFixture()
 
-      await expect(nameWrapper)
-        .write('setSubnodeRecord', [
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
           namehash(name),
           sublabel,
           zeroAddress,
@@ -112,17 +114,15 @@ export const setSubnodeRecordTests = () =>
           0n,
           0,
           0n,
-        ])
-        .toBeRevertedWithString('ERC1155: mint to the zero address')
+        ]),
+      ).toBeRevertedWithString('ERC1155: mint to the zero address')
     })
 
     it('Will not allow wrapping with a target address of the wrapper contract address.', async () => {
-      const { nameWrapper, resolverAddress } = await loadFixture(
-        setSubnodeRecordFixture,
-      )
+      const { nameWrapper, resolverAddress } = await loadFixture()
 
-      await expect(nameWrapper)
-        .write('setSubnodeRecord', [
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
           namehash(name),
           sublabel,
           nameWrapper.address,
@@ -130,23 +130,22 @@ export const setSubnodeRecordTests = () =>
           0n,
           0,
           0n,
-        ])
-        .toBeRevertedWithString(
-          'ERC1155: newOwner cannot be the NameWrapper contract',
-        )
+        ]),
+      ).toBeRevertedWithString(
+        'ERC1155: newOwner cannot be the NameWrapper contract',
+      )
     })
 
     it('Does not allow anyone else to wrap a name even if the owner has authorised the wrapper with the ENS registry.', async () => {
       const { ensRegistry, nameWrapper, accounts, resolverAddress } =
-        await loadFixture(setSubnodeRecordFixture)
+        await loadFixture()
 
       await expectOwnerOf(name).on(nameWrapper).toBe(accounts[0])
 
       await ensRegistry.write.setApprovalForAll([accounts[1].address, true])
 
-      await expect(nameWrapper)
-        .write(
-          'setSubnodeRecord',
+      await expect(
+        nameWrapper.write.setSubnodeRecord(
           [
             namehash(name),
             sublabel,
@@ -157,21 +156,22 @@ export const setSubnodeRecordTests = () =>
             0n,
           ],
           { account: accounts[1] },
-        )
+        ),
+      )
         .toBeRevertedWithCustomError('Unauthorised')
-        .withArgs(namehash(name), getAddress(accounts[1].address))
+        .withArgs([namehash(name), getAddress(accounts[1].address)])
     })
 
     it('Does not allow fuses to be burned if PARENT_CANNOT_CONTROL is not burned.', async () => {
-      const { nameWrapper, actions, accounts } = await loadFixture(fixture)
+      const { nameWrapper, actions, accounts } = await loadNameWrapperFixture()
 
       await actions.registerSetupAndWrapName({
         label,
         fuses: CAN_DO_EVERYTHING,
       })
 
-      await expect(nameWrapper)
-        .write('setSubnodeRecord', [
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
           namehash(name),
           sublabel,
           accounts[0].address,
@@ -179,21 +179,22 @@ export const setSubnodeRecordTests = () =>
           0n,
           CANNOT_UNWRAP,
           MAX_EXPIRY,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(namehash(subname))
+        .withArgs([namehash(subname)])
     })
 
     it('Does not allow fuses to be burned if CANNOT_UNWRAP is not burned', async () => {
-      const { nameWrapper, actions, accounts } = await loadFixture(fixture)
+      const { nameWrapper, actions, accounts } = await loadNameWrapperFixture()
 
       await actions.registerSetupAndWrapName({
         label,
         fuses: CAN_DO_EVERYTHING,
       })
 
-      await expect(nameWrapper)
-        .write('setSubnodeRecord', [
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
           namehash(name),
           sublabel,
           accounts[0].address,
@@ -201,13 +202,14 @@ export const setSubnodeRecordTests = () =>
           0n,
           PARENT_CANNOT_CONTROL | CANNOT_TRANSFER,
           MAX_EXPIRY,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(namehash(subname))
+        .withArgs([namehash(subname)])
     })
 
     it('Fuses will remain 0 if expired', async () => {
-      const { nameWrapper, actions, accounts } = await loadFixture(fixture)
+      const { nameWrapper, actions, accounts } = await loadNameWrapperFixture()
 
       await actions.registerSetupAndWrapName({
         label,
@@ -230,7 +232,7 @@ export const setSubnodeRecordTests = () =>
     })
 
     it('Allows fuses to be burned if not expired and PARENT_CANNOT_CONTROL/CANNOT_UNWRAP are burned', async () => {
-      const { nameWrapper, actions, accounts } = await loadFixture(fixture)
+      const { nameWrapper, actions, accounts } = await loadNameWrapperFixture()
 
       await actions.registerSetupAndWrapName({
         label,
@@ -255,15 +257,15 @@ export const setSubnodeRecordTests = () =>
     })
 
     it('does not allow burning IS_DOT_ETH', async () => {
-      const { nameWrapper, actions, accounts } = await loadFixture(fixture)
+      const { nameWrapper, actions, accounts } = await loadNameWrapperFixture()
 
       await actions.registerSetupAndWrapName({
         label,
         fuses: CANNOT_UNWRAP,
       })
 
-      await expect(nameWrapper)
-        .write('setSubnodeRecord', [
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
           namehash(name),
           sublabel,
           accounts[0].address,
@@ -271,18 +273,17 @@ export const setSubnodeRecordTests = () =>
           0n,
           PARENT_CANNOT_CONTROL | CANNOT_UNWRAP | CANNOT_TRANSFER | IS_DOT_ETH,
           MAX_EXPIRY,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(namehash(subname))
+        .withArgs([namehash(subname)])
     })
 
     it('Emits Wrap event', async () => {
-      const { nameWrapper, accounts } = await loadFixture(
-        setSubnodeRecordFixture,
-      )
+      const { nameWrapper, accounts } = await loadFixture()
 
-      await expect(nameWrapper)
-        .write('setSubnodeRecord', [
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
           namehash(name),
           sublabel,
           accounts[1].address,
@@ -290,24 +291,23 @@ export const setSubnodeRecordTests = () =>
           0n,
           0,
           0n,
-        ])
+        ]),
+      )
         .toEmitEvent('NameWrapped')
-        .withArgs(
-          namehash(subname),
-          dnsEncodeName(subname),
-          accounts[1].address,
-          0,
-          0n,
-        )
+        .withArgs({
+          node: namehash(subname),
+          name: dnsEncodeName(subname),
+          owner: getAddress(accounts[1].address),
+          fuses: 0,
+          expiry: 0n,
+        })
     })
 
     it('Emits TransferSingle event', async () => {
-      const { nameWrapper, accounts } = await loadFixture(
-        setSubnodeRecordFixture,
-      )
+      const { nameWrapper, accounts } = await loadFixture()
 
-      await expect(nameWrapper)
-        .write('setSubnodeRecord', [
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
           namehash(name),
           sublabel,
           accounts[1].address,
@@ -315,21 +315,21 @@ export const setSubnodeRecordTests = () =>
           0n,
           0,
           0n,
-        ])
+        ]),
+      )
         .toEmitEvent('TransferSingle')
-        .withArgs(
-          accounts[0].address,
-          zeroAddress,
-          accounts[1].address,
-          toNameId(subname),
-          1n,
-        )
+        .withArgs({
+          operator: getAddress(accounts[0].address),
+          from: zeroAddress,
+          to: getAddress(accounts[1].address),
+          id: toNameId(subname),
+          value: 1n,
+        })
     })
 
     it('Sets the appropriate values on the ENS registry', async () => {
-      const { ensRegistry, nameWrapper, actions, accounts } = await loadFixture(
-        setSubnodeRecordFixture,
-      )
+      const { ensRegistry, nameWrapper, actions, accounts } =
+        await loadFixture()
 
       await actions.setSubnodeRecord.onNameWrapper({
         parentName: name,
@@ -351,12 +351,10 @@ export const setSubnodeRecordTests = () =>
     })
 
     it('Will not create a subdomain with an empty label', async () => {
-      const { nameWrapper, resolverAddress } = await loadFixture(
-        setSubnodeRecordFixture,
-      )
+      const { nameWrapper, resolverAddress } = await loadFixture()
 
-      await expect(nameWrapper)
-        .write('setSubnodeRecord', [
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
           namehash(name),
           '',
           zeroAddress,
@@ -364,13 +362,13 @@ export const setSubnodeRecordTests = () =>
           0n,
           0,
           0n,
-        ])
-        .toBeRevertedWithCustomError('LabelTooShort')
+        ]),
+      ).toBeRevertedWithCustomError('LabelTooShort')
     })
 
     it('should be able to call twice and change the owner', async () => {
       const { nameWrapper, actions, accounts, resolverAddress } =
-        await loadFixture(setSubnodeRecordFixture)
+        await loadFixture()
 
       await actions.setSubnodeRecord.onNameWrapper({
         parentName: name,
@@ -398,7 +396,7 @@ export const setSubnodeRecordTests = () =>
     })
 
     it('setting owner to 0 burns and unwraps', async () => {
-      const { nameWrapper, actions, accounts } = await loadFixture(fixture)
+      const { nameWrapper, actions, accounts } = await loadNameWrapperFixture()
 
       await actions.registerSetupAndWrapName({
         label,
@@ -421,8 +419,8 @@ export const setSubnodeRecordTests = () =>
 
       await expectOwnerOf(subname).on(nameWrapper).toBe(accounts[1])
 
-      await expect(nameWrapper)
-        .write('setSubnodeRecord', [
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
           namehash(name),
           sublabel,
           zeroAddress,
@@ -430,15 +428,16 @@ export const setSubnodeRecordTests = () =>
           0n,
           PARENT_CANNOT_CONTROL,
           MAX_EXPIRY,
-        ])
+        ]),
+      )
         .toEmitEvent('NameUnwrapped')
-        .withArgs(namehash(subname), zeroAddress)
+        .withArgs({ node: namehash(subname), owner: zeroAddress })
 
       await expectOwnerOf(subname).on(nameWrapper).toBe(zeroAccount)
     })
 
     it('Unwrapping within an external contract does not create any state inconsistencies', async () => {
-      const { nameWrapper, accounts, actions } = await loadFixture(fixture)
+      const { nameWrapper, accounts, actions } = await loadNameWrapperFixture()
 
       await actions.setRegistryApprovalForWrapper()
       await actions.registerSetupAndWrapName({
@@ -446,7 +445,7 @@ export const setSubnodeRecordTests = () =>
         fuses: CAN_DO_EVERYTHING,
       })
 
-      const testReentrancy = await hre.viem.deployContract(
+      const testReentrancy = await connection.viem.deployContract(
         'TestNameWrapperReentrancy',
         [
           accounts[0].address,
@@ -469,8 +468,8 @@ export const setSubnodeRecordTests = () =>
       })
 
       // move owner to testReentrancy, which unwraps domain itself to account while keeping ERC1155 to testReentrancy
-      await expect(nameWrapper)
-        .write('setSubnodeRecord', [
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
           namehash(name),
           sublabel,
           testReentrancy.address,
@@ -478,16 +477,17 @@ export const setSubnodeRecordTests = () =>
           0n,
           CANNOT_UNWRAP | PARENT_CANNOT_CONTROL,
           MAX_EXPIRY,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(namehash(subname))
+        .withArgs([namehash(subname)])
 
       // reverts because CANNOT_UNWRAP/PCC are burned first, and then unwrap is attempted inside contract, which fails, because CU has already been burned
     })
 
     it('Unwrapping a previously wrapped unexpired name retains PCC and so reverts setSubnodeRecord', async () => {
-      const { ensRegistry, nameWrapper, actions, accounts, baseRegistrar } =
-        await loadFixture(fixture)
+      const { nameWrapper, actions, accounts, baseRegistrar } =
+        await loadNameWrapperFixture()
 
       await actions.registerSetupAndWrapName({
         label,
@@ -534,8 +534,8 @@ export const setSubnodeRecordTests = () =>
       expect(expiry).toEqual(parentExpiry + GRACE_PERIOD)
 
       // attempt to rewrap with PCC still burnt
-      await expect(nameWrapper)
-        .write('setSubnodeRecord', [
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
           namehash(name),
           sublabel,
           accounts[1].address,
@@ -543,14 +543,15 @@ export const setSubnodeRecordTests = () =>
           0n,
           0,
           0n,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(namehash(subname))
+        .withArgs([namehash(subname)])
     })
 
     it('Rewrapping a name that had PCC burned, but has now expired is possible', async () => {
       const { nameWrapper, actions, accounts, baseRegistrar, testClient } =
-        await loadFixture(fixture)
+        await loadNameWrapperFixture()
 
       await actions.registerSetupAndWrapName({
         label,
@@ -625,7 +626,7 @@ export const setSubnodeRecordTests = () =>
         baseRegistrar,
         testClient,
         publicClient,
-      } = await loadFixture(fixture)
+      } = await loadNameWrapperFixture()
 
       const sublabel2 = 'sub2'
 
@@ -657,8 +658,8 @@ export const setSubnodeRecordTests = () =>
         CANNOT_CREATE_SUBDOMAIN,
       ])
 
-      await expect(nameWrapper)
-        .write('setSubnodeRecord', [
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
           namehash(name),
           sublabel2,
           accounts[1].address,
@@ -666,9 +667,10 @@ export const setSubnodeRecordTests = () =>
           0n,
           0,
           parentExpiry - DAY / 2n,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(namehash(`${sublabel2}.${name}`))
+        .withArgs([namehash(`${sublabel2}.${name}`)])
 
       await testClient.increaseTime({ seconds: Number(DAY / 2n + 1n) })
       await testClient.mine({ blocks: 1 })
@@ -684,8 +686,8 @@ export const setSubnodeRecordTests = () =>
       expect(fuses).toEqual(0)
       expect(expiry).toBeLessThan(timestamp)
 
-      await expect(nameWrapper)
-        .write('setSubnodeRecord', [
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
           namehash(name),
           sublabel,
           accounts[1].address,
@@ -693,9 +695,10 @@ export const setSubnodeRecordTests = () =>
           0n,
           0,
           parentExpiry - DAY / 2n,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(namehash(subname))
+        .withArgs([namehash(subname)])
     })
 
     it('Burning a name still protects it from the parent as long as it is unexpired and has PCC burnt', async () => {
@@ -706,7 +709,7 @@ export const setSubnodeRecordTests = () =>
         accounts,
         baseRegistrar,
         publicClient,
-      } = await loadFixture(fixture)
+      } = await loadNameWrapperFixture()
 
       await actions.registerSetupAndWrapName({
         label,
@@ -761,8 +764,8 @@ export const setSubnodeRecordTests = () =>
       await expectOwnerOf(subname).on(ensRegistry).toBe(zeroAccount)
 
       // attempt to take back the name
-      await expect(nameWrapper)
-        .write('setSubnodeRecord', [
+      await expect(
+        nameWrapper.write.setSubnodeRecord([
           namehash(name),
           sublabel,
           accounts[0].address,
@@ -770,8 +773,9 @@ export const setSubnodeRecordTests = () =>
           0n,
           PARENT_CANNOT_CONTROL,
           MAX_EXPIRY,
-        ])
+        ]),
+      )
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(namehash(subname))
+        .withArgs([namehash(subname)])
     })
   })
