@@ -1,20 +1,23 @@
-import { loadFixture } from '@nomicfoundation/hardhat-toolbox-viem/network-helpers.js'
-import { expect } from 'chai'
+import type { NetworkConnection } from 'hardhat/types/network'
 import { getAddress, namehash } from 'viem'
+
 import {
   CANNOT_SET_TTL,
   CANNOT_UNWRAP,
   expectOwnerOf,
-  deployNameWrapperWithUtils as fixture,
+  type LoadNameWrapperFixture,
 } from '../fixtures/utils.js'
 
-export const setTTLTests = () =>
+export const setTTLTests = (
+  connection: NetworkConnection,
+  loadNameWrapperFixture: LoadNameWrapperFixture,
+) =>
   describe('setTTL', () => {
     const label = 'setttl'
     const name = `${label}.eth`
 
-    async function setTTLFixture() {
-      const initial = await loadFixture(fixture)
+    async function fixture() {
+      const initial = await loadNameWrapperFixture()
       const { actions } = initial
 
       await actions.registerSetupAndWrapName({
@@ -24,9 +27,11 @@ export const setTTLTests = () =>
 
       return initial
     }
+    const loadFixture = async () =>
+      connection.networkHelpers.loadFixture(fixture)
 
     it('Can be called by the owner', async () => {
-      const { nameWrapper, accounts } = await loadFixture(setTTLFixture)
+      const { nameWrapper, accounts } = await loadFixture()
 
       await expectOwnerOf(name).on(nameWrapper).toBe(accounts[0])
 
@@ -34,7 +39,7 @@ export const setTTLTests = () =>
     })
 
     it('Performs the appropriate function on the ENS registry.', async () => {
-      const { ensRegistry, nameWrapper } = await loadFixture(setTTLFixture)
+      const { ensRegistry, nameWrapper } = await loadFixture()
 
       await expect(ensRegistry.read.ttl([namehash(name)])).resolves.toEqual(0n)
 
@@ -46,7 +51,7 @@ export const setTTLTests = () =>
     })
 
     it('Can be called by an account authorised by the owner.', async () => {
-      const { nameWrapper, accounts } = await loadFixture(setTTLFixture)
+      const { nameWrapper, accounts } = await loadFixture()
 
       await expectOwnerOf(name).on(nameWrapper).toBe(accounts[0])
 
@@ -58,22 +63,24 @@ export const setTTLTests = () =>
     })
 
     it('Cannot be called by anyone else.', async () => {
-      const { nameWrapper, accounts } = await loadFixture(setTTLFixture)
+      const { nameWrapper, accounts } = await loadFixture()
 
-      await expect(nameWrapper)
-        .write('setTTL', [namehash(name), 3600n], { account: accounts[1] })
+      await expect(
+        nameWrapper.write.setTTL([namehash(name), 3600n], {
+          account: accounts[1],
+        }),
+      )
         .toBeRevertedWithCustomError('Unauthorised')
-        .withArgs(namehash(name), getAddress(accounts[1].address))
+        .withArgs([namehash(name), getAddress(accounts[1].address)])
     })
 
     it('Cannot be called if CANNOT_SET_TTL is burned', async () => {
-      const { nameWrapper } = await loadFixture(setTTLFixture)
+      const { nameWrapper } = await loadFixture()
 
       await nameWrapper.write.setFuses([namehash(name), CANNOT_SET_TTL])
 
-      await expect(nameWrapper)
-        .write('setTTL', [namehash(name), 100n])
+      await expect(nameWrapper.write.setTTL([namehash(name), 100n]))
         .toBeRevertedWithCustomError('OperationProhibited')
-        .withArgs(namehash(name))
+        .withArgs([namehash(name)])
     })
   })
